+++ JupiterSDK/Classes/ServiceManager.swift
@@ -3,7 +3,7 @@ import CoreMotion
 import UIKit
 
 public class ServiceManager: Observation {
-    public static let sdkVersion: String = "3.3.4"
+    public static let sdkVersion: String = "3.3.5"
     
     func tracking(input: FineLocationTrackingResult, isPast: Bool) {
         for observer in observers {
@@ -86,11 +86,6 @@ public class ServiceManager: Observation {
     var os: String = "Unknown"
     var osVersion: Int = 0
     
-    var PathType = [String: [Int]]()
-    var PathPoint = [String: [[Double]]]()
-    var PathPointMinMax = [Double]()
-    var PathMagScale = [String: [Double]]()
-    var PathHeading = [String: [String]]()
     var LoadPathPoint = [String: Bool]()
     
     var EntranceArea = [String: [[Double]]]()
@@ -98,7 +93,6 @@ public class ServiceManager: Observation {
     var EntranceWards = [String: [String: Int]]()
     var allEntranceWards = [String]()
     var LevelChangeArea = [String: [[Double]]]()
-    var EntranceMatchingArea = [String: [[Double]]]()
     var EntranceNumbers: Int = 0
     var EntranceLevelInfo = [String: [String]]()
     var EntranceInfo = [String: [[Double]]]()
@@ -110,7 +104,6 @@ public class ServiceManager: Observation {
     var isPhaseBreakInSimulate: Bool = false
     var isStopReqeust: Bool = false
     var indexAfterSimulate: Int = 0
-    var posBufferForEndSimulate =  [[Double]]()
     var buildingsAndLevels = [String:[String]]()
     
     public var isLoadEnd = [String: [Bool]]()
@@ -135,6 +128,8 @@ public class ServiceManager: Observation {
     // ---------------------------- //
     
     var paramEstimator = ParameterEstimator()
+    var pmCalculator = PathMatchingCalculator()
+    
     // ----- Timer ----- //
     var backgroundUpTimer: DispatchSourceTimer?
     var backgroundUvTimer: DispatchSourceTimer?
@@ -202,11 +197,8 @@ public class ServiceManager: Observation {
     var USER_TRAJECTORY_DIAGONAL: Double = 20
     var NUM_STRAIGHT_INDEX_DR = 10
     var NUM_STRAIGHT_INDEX_PDR = 10
-    var preTailHeading: Double = 0
-    var preTuMmHeading: Double = 0
     var magHeading: Double = 0
     
-    var unitDistane: Double = 0
     var isStartFlag: Bool = false
     var isStartComplete: Bool = false
     var lookingState: Bool = true
@@ -237,8 +229,6 @@ public class ServiceManager: Observation {
     var isGetFirstResponse: Bool = false
     var indexAfterResponse: Int = 0
     
-    var rssiBias: Int = 0
-    
     var isPossibleNormalize: Bool = false
     var deviceMinRss: Double = -100.0
     var standardMinRss: Double = -99.0
@@ -248,17 +238,12 @@ public class ServiceManager: Observation {
     var isScaleLoaded: Bool = false
     var isScaleConverged: Bool = false
     
-    var isBiasConverged: Bool = false
     var sccBadCount: Int = 0
     var scCompensationArray: [Double] = [0.8, 1.0, 1.2]
     var scCompensation: Double = 1.0
     var scCompensationBadCount: Int = 0
     var scVelocityScale: Double = 1.0
     var entranceVelocityScale: Double = 1.0
-    
-    var sccGoodBiasArray = [Int]()
-    var biasRequestTime: Int = 0
-    var isBiasRequested: Bool = false
     
     var scRequestTime: Int = 0
     var isScRequested: Bool = false
@@ -276,13 +261,11 @@ public class ServiceManager: Observation {
     var indexPast: Int = 0
     
     var indexSend: Int = 0
-    var indexReceived: Int = 0
     
     var timeUpdateFlag: Bool = false
     var measurementUpdateFlag: Bool = false
     var isPhaseBreak: Bool = false
     var isNeedTrajInit: Bool = false
-    var indexAfterTrajInit: Int = 0
     
     var kalmanP: Double = 1
     var kalmanQ: Double = 0.3
@@ -335,7 +318,6 @@ public class ServiceManager: Observation {
     var timeFailRF: Double = 0
     var timeActiveUV: Double = 0
     var timeRequest: Double = 0
-    var timePhaseChange: Double = 0
     var timeSleepRF: Double = 0
     var timeSleepUV: Double = 0
     
@@ -640,13 +622,13 @@ public class ServiceManager: Observation {
                                                 let url = self.getPpUrl(server: self.serverType, key: key)
                                                 let urlComponents = URLComponents(string: url)
                                                 let requestURL = URLRequest(url: (urlComponents?.url)!)
-                                                let dataTask = URLSession.shared.dataTask(with: requestURL, completionHandler: { (data, response, error) in
+                                                let dataTask = URLSession.shared.dataTask(with: requestURL, completionHandler: { [self] (data, response, error) in
                                                     let statusCode = (response as? HTTPURLResponse)?.statusCode ?? 500
                                                     
                                                     if (statusCode == 200) {
                                                         if let responseData = data {
                                                             if let utf8Text = String(data: responseData, encoding: .utf8) {
-                                                                ( self.PathType[key], self.PathPoint[key], self.PathMagScale[key], self.PathHeading[key] ) = self.parseRoad(data: utf8Text)
+                                                                ( pmCalculator.PathType[key], pmCalculator.PathPoint[key], pmCalculator.PathMagScale[key], pmCalculator.PathHeading[key] ) = pmCalculator.parseRoad(data: utf8Text)
                                                                 self.isLoadEnd[key] = [true, true]
                                                             }
                                                         }
@@ -682,7 +664,7 @@ public class ServiceManager: Observation {
                                                         let result = decodeGeo(json: returnedString)
                                                         let key: String = "\(buildingGeo)_\(levelGeo)"
                                                         self.EntranceArea[key] = result.entrance_area
-                                                        self.EntranceMatchingArea[key] = result.entrance_matching_area
+                                                        pmCalculator.EntranceMatchingArea[key] = result.entrance_matching_area
                                                         self.LevelChangeArea[key] = result.level_change_area
                                                         
                                                         countBuildingLevel += 1
@@ -715,10 +697,7 @@ public class ServiceManager: Observation {
                                                                                             let result = decodeParam(json: returnedString)
                                                                                             if (result.rss_compensations.isEmpty) {
                                                                                                 // Need Bias Estimation
-                                                                                                displayOutput.bias = self.rssiBias
-                                                                                                displayOutput.isConverged = self.isBiasConverged
                                                                                                 self.isScaleLoaded = false
-                                                                                                
                                                                                                 self.isStartComplete = true
                                                                                                 self.startTimer()
                                                                                                 
@@ -741,11 +720,7 @@ public class ServiceManager: Observation {
                                                                                                         self.preNormalizationScale = paramFromServer.normalization_scale
                                                                                                         print(localTime + " , (Jupiter) Load Param (Device) : \(paramFromServer.normalization_scale)")
                                                                                                     }
-                                                                                                    
-                                                                                                    displayOutput.bias = self.rssiBias
-                                                                                                    displayOutput.isConverged = self.isBiasConverged
                                                                                                     self.isScaleLoaded = true
-                                                                                                    
                                                                                                     self.isStartComplete = true
                                                                                                     self.startTimer()
                                                                                                     
@@ -754,10 +729,7 @@ public class ServiceManager: Observation {
                                                                                                     self.reporting(input: START_FLAG)
                                                                                                     completion(true, message)
                                                                                                 } else {
-                                                                                                    displayOutput.bias = self.rssiBias
-                                                                                                    displayOutput.isConverged = self.isBiasConverged
                                                                                                     self.isScaleLoaded = false
-                                                                                                    
                                                                                                     self.isStartComplete = true
                                                                                                     self.startTimer()
                                                                                                     
@@ -789,11 +761,7 @@ public class ServiceManager: Observation {
                                                                                         self.preNormalizationScale = paramFromServer.normalization_scale
                                                                                         print(localTime + " , (Jupiter) Load Param (Device) : \(paramFromServer.normalization_scale)")
                                                                                     }
-                                                                                    
-                                                                                    displayOutput.bias = self.rssiBias
-                                                                                    displayOutput.isConverged = self.isBiasConverged
                                                                                     self.isScaleLoaded = true
-                                                                                    
                                                                                     self.isStartComplete = true
                                                                                     self.startTimer()
                                                                                     
@@ -886,23 +854,6 @@ public class ServiceManager: Observation {
             }
         }
     }
-    
-    func findClosestStructure(to myOsVersion: Int, in array: [rss_compensation]) -> rss_compensation? {
-        guard let first = array.first else {
-            return nil
-        }
-        var closest = first
-        var closestDistance = closest.os_version - myOsVersion
-        for d in array {
-            let distance = d.os_version - myOsVersion
-            if abs(distance) < abs(closestDistance) {
-                closest = d
-                closestDistance = distance
-            }
-        }
-        return closest
-    }
-
     
     public func setSaveFlag(flag: Bool) {
         self.isSaveFlag = flag
@@ -1143,7 +1094,6 @@ public class ServiceManager: Observation {
         self.isPhaseBreak = false
         self.isGetFirstResponse = false
         self.isNeedTrajInit = false
-        self.indexAfterTrajInit = 0
         
         self.isActiveKf = false
         self.updateHeading = 0
@@ -1289,11 +1239,31 @@ public class ServiceManager: Observation {
         }
     }
     
-    public func getRecentResult(id: String, completion: @escaping (Int, String) -> Void) {
+    public func getRecentResult(id: String, completion: @escaping (Int, FineLocationTrackingFromServer) -> Void) {
         let currentTime: Int = getCurrentTimeInMilliseconds()
         let input = RecentResult(user_id: id, mobile_time: currentTime)
-        NetworkManager.shared.postRecent(url: RECENT_URL, input: input, completion: { statusCode, returnedString in
-            completion(statusCode, returnedString)
+        NetworkManager.shared.postRecent(url: RECENT_URL, input: input, completion: { [self] statusCode, returnedString in
+            if (statusCode == 200) {
+                let result = jsonToResult(json: returnedString)
+                var recentResult = result
+                
+                let pathMatchingResult = pmCalculator.pathMatching(building: result.building_name, level: result.level_name, x: result.x, y: result.y, heading: result.absolute_heading, isPast: false, HEADING_RANGE: 46, isUseHeading: true, pathType: 0, range: 10)
+                let resultTime: Int = result.mobile_time
+                let resultIndex = result.index
+                let resultBuildingName: String = result.building_name
+                let resultLevelNameAll: String = result.level_name
+                let resultLevelName: String = removeLevelDirectionString(levelName: resultLevelNameAll)
+                
+                recentResult.building_name = resultBuildingName
+                recentResult.level_name = resultLevelName
+                recentResult.x = pathMatchingResult.xyhs[0]
+                recentResult.y = pathMatchingResult.xyhs[1]
+                recentResult.absolute_heading = pathMatchingResult.xyhs[2]
+                
+                completion(statusCode, recentResult)
+            } else {
+                completion(statusCode, FineLocationTrackingFromServer())
+            }
         })
     }
     
@@ -1308,17 +1278,17 @@ public class ServiceManager: Observation {
                 recentResult.building = decodedResult.building_name
                 recentResult.level = decodedResult.level_name
                 
-                let pmResultWithHeading = pathMatching(building: decodedResult.building_name, level: decodedResult.level_name, x: decodedResult.x, y: decodedResult.y, heading: decodedResult.absolute_heading, tuXY: [0, 0], isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: true, pathType: 1)
+                let pmResultWithHeading = pmCalculator.pathMatching(building: decodedResult.building_name, level: decodedResult.level_name, x: decodedResult.x, y: decodedResult.y, heading: decodedResult.absolute_heading, isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: true, pathType: 1, range: SQUARE_RANGE)
                 
                 if (pmResultWithHeading.isSuccess) {
-                    recentResult.x = pmResultWithHeading.xyh[0]
-                    recentResult.y = pmResultWithHeading.xyh[1]
-                    recentResult.heading = pmResultWithHeading.xyh[2]
+                    recentResult.x = pmResultWithHeading.xyhs[0]
+                    recentResult.y = pmResultWithHeading.xyhs[1]
+                    recentResult.heading = pmResultWithHeading.xyhs[2]
                 } else {
-                    let pmResult = pathMatching(building: decodedResult.building_name, level: decodedResult.level_name, x: decodedResult.x, y: decodedResult.y, heading: decodedResult.absolute_heading, tuXY: [0, 0], isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: false, pathType: 1)
-                    recentResult.x = pmResult.xyh[0]
-                    recentResult.y = pmResult.xyh[1]
-                    recentResult.heading = pmResult.xyh[2]
+                    let pmResult = pmCalculator.pathMatching(building: decodedResult.building_name, level: decodedResult.level_name, x: decodedResult.x, y: decodedResult.y, heading: decodedResult.absolute_heading, isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: false, pathType: 1, range: SQUARE_RANGE)
+                    recentResult.x = pmResult.xyhs[0]
+                    recentResult.y = pmResult.xyhs[1]
+                    recentResult.heading = pmResult.xyhs[2]
                 }
                 recentResult.isIndoor = true
                 
@@ -1489,12 +1459,12 @@ public class ServiceManager: Observation {
                 self.headingBeforePm = result.absolute_heading
                 if (runMode == "pdr") {
                     let isUseHeading: Bool = false
-                    let correctResult = pathMatching(building: buildingName, level: levelName, x: result.x, y: result.y, heading: result.absolute_heading, tuXY: [0,0], isPast: isPast, HEADING_RANGE: HEADING_RANGE, isUseHeading: false, pathType: 0)
+                    let correctResult = pmCalculator.pathMatching(building: buildingName, level: levelName, x: result.x, y: result.y, heading: result.absolute_heading, isPast: isPast, HEADING_RANGE: HEADING_RANGE, isUseHeading: isUseHeading, pathType: 0, range: SQUARE_RANGE)
                     if (correctResult.isSuccess) {
                         displayOutput.isPmSuccess = true
-                        result.x = correctResult.xyh[0]
-                        result.y = correctResult.xyh[1]
-                        result.absolute_heading = correctResult.xyh[2]
+                        result.x = correctResult.xyhs[0]
+                        result.y = correctResult.xyhs[1]
+                        result.absolute_heading = correctResult.xyhs[2]
                     } else {
                         displayOutput.isPmSuccess = false
                         let localTime: String = getLocalTimeString()
@@ -1503,7 +1473,7 @@ public class ServiceManager: Observation {
                         
                         var isLoadPathPoint: Bool = true
                         if let isLoad: Bool = self.LoadPathPoint[key] { isLoadPathPoint = isLoad }
-                        if let mainRoad: [[Double]] = self.PathPoint[key] {
+                        if let mainRoad: [[Double]] = pmCalculator.PathPoint[key] {
                             self.LoadPathPoint[key] = true
                         } else {
                             if (isLoadPathPoint) {
@@ -1512,12 +1482,12 @@ public class ServiceManager: Observation {
                                 
                                 let urlComponents = URLComponents(string: url)
                                 let requestURL = URLRequest(url: (urlComponents?.url)!)
-                                let dataTask = URLSession.shared.dataTask(with: requestURL, completionHandler: { (data, response, error) in
+                                let dataTask = URLSession.shared.dataTask(with: requestURL, completionHandler: { [self] (data, response, error) in
                                     let statusCode = (response as? HTTPURLResponse)?.statusCode ?? 500
                                     if (statusCode == 200) {
                                         if let responseData = data {
                                             if let utf8Text = String(data: responseData, encoding: .utf8) {
-                                                ( self.PathType[key], self.PathPoint[key], self.PathMagScale[key], self.PathHeading[key] ) = self.parseRoad(data: utf8Text)
+                                                ( pmCalculator.PathType[key], pmCalculator.PathPoint[key], pmCalculator.PathMagScale[key], pmCalculator.PathHeading[key] ) = pmCalculator.parseRoad(data: utf8Text)
                                                 self.LoadPathPoint[key] = true
                                                 let log: String = localTime + " , (Jupiter) Success : Load \(buildingName) \(levelName) Path-Point (when PP was empty)"
                                                 print(log)
@@ -1536,11 +1506,11 @@ public class ServiceManager: Observation {
                         if (self.isActiveKf) {
                             result = self.lastResult
                         } else {
-                            let correctResult = pathMatching(building: buildingName, level: levelName, x: result.x, y: result.y, heading: result.absolute_heading, tuXY: [0,0], isPast: isPast, HEADING_RANGE: HEADING_RANGE, isUseHeading: false, pathType: 0)
+                            let correctResult = pmCalculator.pathMatching(building: buildingName, level: levelName, x: result.x, y: result.y, heading: result.absolute_heading, isPast: isPast, HEADING_RANGE: HEADING_RANGE, isUseHeading: isUseHeading, pathType: 0, range: SQUARE_RANGE)
                             if (correctResult.isSuccess) {
-                                result.x = correctResult.xyh[0]
-                                result.y = correctResult.xyh[1]
-                                result.absolute_heading = correctResult.xyh[2]
+                                result.x = correctResult.xyhs[0]
+                                result.y = correctResult.xyhs[1]
+                                result.absolute_heading = correctResult.xyhs[2]
                             }
                         }
                     }
@@ -1549,12 +1519,13 @@ public class ServiceManager: Observation {
                     if (isVenusMode) {
                         isUseHeading = false
                     }
-                    let correctResult = pathMatching(building: buildingName, level: levelName, x: result.x, y: result.y, heading: result.absolute_heading, tuXY: [0,0], isPast: isPast, HEADING_RANGE: HEADING_RANGE, isUseHeading: isUseHeading, pathType: 1)
+                    let correctResult = pmCalculator.pathMatching(building: buildingName, level: levelName, x: result.x, y: result.y, heading: result.absolute_heading, isPast: isPast, HEADING_RANGE: HEADING_RANGE, isUseHeading: isUseHeading, pathType: 1, range: SQUARE_RANGE)
+                    unitDRGenerator.setVelocityScaleFactor(scaleFactor: correctResult.xyhs[3])
                     if (correctResult.isSuccess) {
                         displayOutput.isPmSuccess = true
-                        result.x = correctResult.xyh[0]
-                        result.y = correctResult.xyh[1]
-                        result.absolute_heading = correctResult.xyh[2]
+                        result.x = correctResult.xyhs[0]
+                        result.y = correctResult.xyhs[1]
+                        result.absolute_heading = correctResult.xyhs[2]
                     } else {
                         displayOutput.isPmSuccess = false
                         let localTime: String = getLocalTimeString()
@@ -1563,7 +1534,7 @@ public class ServiceManager: Observation {
                         
                         var isLoadPathPoint: Bool = true
                         if let isLoad: Bool = self.LoadPathPoint[key] { isLoadPathPoint = isLoad }
-                        if let mainRoad: [[Double]] = self.PathPoint[key] {
+                        if let mainRoad: [[Double]] = pmCalculator.PathPoint[key] {
                             self.LoadPathPoint[key] = true
                         } else {
                             if (isLoadPathPoint) {
@@ -1572,12 +1543,12 @@ public class ServiceManager: Observation {
                                 
                                 let urlComponents = URLComponents(string: url)
                                 let requestURL = URLRequest(url: (urlComponents?.url)!)
-                                let dataTask = URLSession.shared.dataTask(with: requestURL, completionHandler: { (data, response, error) in
+                                let dataTask = URLSession.shared.dataTask(with: requestURL, completionHandler: { [self] (data, response, error) in
                                     let statusCode = (response as? HTTPURLResponse)?.statusCode ?? 500
                                     if (statusCode == 200) {
                                         if let responseData = data {
                                             if let utf8Text = String(data: responseData, encoding: .utf8) {
-                                                ( self.PathType[key], self.PathPoint[key], self.PathMagScale[key], self.PathHeading[key] ) = self.parseRoad(data: utf8Text)
+                                                ( pmCalculator.PathType[key], pmCalculator.PathPoint[key], pmCalculator.PathMagScale[key], pmCalculator.PathHeading[key] ) = pmCalculator.parseRoad(data: utf8Text)
                                                 self.LoadPathPoint[key] = true
                                                 let log: String = localTime + " , (Jupiter) Success : Load \(buildingName) \(levelName) Path-Point (when PP was empty)"
                                                 print(log)
@@ -1595,15 +1566,9 @@ public class ServiceManager: Observation {
                         if (self.isActiveKf) {
                             result = self.lastResult
                         } else {
-                            result.x = correctResult.xyh[0]
-                            result.y = correctResult.xyh[1]
-                            result.absolute_heading = correctResult.xyh[2]
-//                            let correctResult = pathMatching(building: buildingName, level: levelName, x: result.x, y: result.y, heading: result.absolute_heading, tuXY: [0,0], isPast: isPast, HEADING_RANGE: HEADING_RANGE, isUseHeading: false, pathType: 1)
-//                            if (correctResult.isSuccess) {
-//                                result.x = correctResult.xyh[0]
-//                                result.y = correctResult.xyh[1]
-//                                result.absolute_heading = correctResult.xyh[2]
-//                            }
+                            result.x = correctResult.xyhs[0]
+                            result.y = correctResult.xyhs[1]
+                            result.absolute_heading = correctResult.xyhs[2]
                         }
                     }
                 }
@@ -1624,8 +1589,6 @@ public class ServiceManager: Observation {
         }
         
         displayOutput.mode = runMode
-        displayOutput.bias = self.rssiBias
-        displayOutput.isConverged = self.isBiasConverged
        
         return result
     }
@@ -1666,7 +1629,7 @@ public class ServiceManager: Observation {
                     }
                     
                     if (!self.isGetFirstResponse) {
-                        let findResult = findNetworkBadEntrance(bleAvg: self.bleAvg, bias: self.rssiBias)
+                        let findResult = findNetworkBadEntrance(bleAvg: self.bleAvg)
                         self.isInNetworkBadEntrance = findResult.0
                         
                         if (!self.isIndoor && (self.timeForInit >= TIME_INIT_THRESHOLD)) {
@@ -1997,7 +1960,9 @@ public class ServiceManager: Observation {
 
                 
                 // Kalman Filter
-                let diffHeading = unitDRInfo.heading - preUnitHeading
+                let diffHeading = unitDRInfo.heading - self.preUnitHeading
+                self.preUnitHeading = unitDRInfo.heading
+                
                 // Time Update
                 if (self.isActiveKf) {
                     if (self.timeUpdateFlag) {
@@ -2110,8 +2075,6 @@ public class ServiceManager: Observation {
                         self.resultToReturn = self.makeOutputResult(input: self.outputResult, isPast: self.flagPast, runMode: self.runMode, isVenusMode: self.isVenusMode)
                     }
                 }
-                
-                preUnitHeading = unitDRInfo.heading
                 
                 // Put UV
                 if ((inputUserVelocity.count-1) >= UVD_INPUT_NUM) {
@@ -2534,7 +2497,7 @@ public class ServiceManager: Observation {
                         diffHeading = 5
                     }
                     // Search Direction
-                    let ppHeadings = getPathMatchingHeadings(building: userBuilding, level: userLevel, x: userX, y: userY, heading: userH, RANGE: RANGE, mode: mode)
+                    let ppHeadings = pmCalculator.getPathMatchingHeadings(building: userBuilding, level: userLevel, x: userX, y: userY, heading: userH, RANGE: RANGE, mode: mode)
                     var searchHeadings: [Double] = []
                     for i in 0..<ppHeadings.count {
                         searchHeadings.append(compensateHeading(heading: ppHeadings[i]-diffHeading))
@@ -2571,7 +2534,6 @@ public class ServiceManager: Observation {
                     
                     tailIndex = userTrajectory[0].index
                     let isStraight = isTrajectoryStraight(for: uvHeading, size: uvHeading.count, mode: mode, conditionPdr: NUM_STRAIGHT_INDEX_PDR, conditionDr: NUM_STRAIGHT_INDEX_DR)
-            
                     let closestIndex = findClosestValueIndex(to: tailIndex, in: pastTrajIndex)
                     if let headingIndex = closestIndex {
 //                        if (isStraight == 1) {
@@ -2613,7 +2575,7 @@ public class ServiceManager: Observation {
                         let headingStart = compensateHeading(heading: headingFromHead[headingFromHead.count-1]-180)
                         let headingEnd = compensateHeading(heading: headingFromHead[0]-180)
 
-                        let areaMinMax: [Double] = getSearchAreaMinMax(xyMinMax: xyMinMax, heading: [headingStart, headingEnd], recentScc: recentScc, searchType: isStraight, mode: mode)
+                        let areaMinMax: [Double] = getSearchAreaMinMax(xyMinMax: xyMinMax, heading: [headingStart, headingEnd], recentScc: recentScc, searchType: isStraight, lengthCondition: USER_TRAJECTORY_DIAGONAL)
                         let searchArea = getSearchCoordinates(areaMinMax: areaMinMax, interval: 1.0)
                         resultRange = areaMinMax.map { Int($0) }
                         
@@ -2655,7 +2617,7 @@ public class ServiceManager: Observation {
                         let headingStart = compensateHeading(heading: headingFromHead[headingFromHead.count-1]-180)
                         let headingEnd = compensateHeading(heading: headingFromHead[0]-180)
                         
-                        let areaMinMax: [Double] = getSearchAreaMinMax(xyMinMax: xyMinMax, heading: [headingStart, headingEnd], recentScc: recentScc, searchType: isStraight, mode: mode)
+                        let areaMinMax: [Double] = getSearchAreaMinMax(xyMinMax: xyMinMax, heading: [headingStart, headingEnd], recentScc: recentScc, searchType: isStraight, lengthCondition: USER_TRAJECTORY_DIAGONAL)
                         let searchArea = getSearchCoordinates(areaMinMax: areaMinMax, interval: 1.0)
                         resultRange = areaMinMax.map { Int($0) }
                         
@@ -2751,7 +2713,7 @@ public class ServiceManager: Observation {
                     
                     
                     // Add
-                    let ppHeadings = getPathMatchingHeadings(building: userBuilding, level: userLevel, x: userX, y: userY, heading: userH, RANGE: RANGE, mode: mode)
+                    let ppHeadings = pmCalculator.getPathMatchingHeadings(building: userBuilding, level: userLevel, x: userX, y: userY, heading: userH, RANGE: RANGE, mode: mode)
                     var searchHeadings: [Double] = []
                     if (accumulatedValue <= 30) {
                         searchHeadings = ppHeadings
@@ -2923,7 +2885,7 @@ public class ServiceManager: Observation {
                     }
                     
                     // Search Direction
-                    let ppHeadings = getPathMatchingHeadings(building: userBuilding, level: userLevel, x: userX, y: userY, heading: userH, RANGE: RANGE, mode: mode)
+                    let ppHeadings = pmCalculator.getPathMatchingHeadings(building: userBuilding, level: userLevel, x: userX, y: userY, heading: userH, RANGE: RANGE, mode: mode)
                     var searchHeadings: [Double] = []
                     for i in 0..<ppHeadings.count {
                         searchHeadings.append(compensateHeading(heading: ppHeadings[i]-10))
@@ -2976,7 +2938,7 @@ public class ServiceManager: Observation {
                         let headingEnd = compensateHeading(heading: headingFromHead[0]-180)
                         let diffHeading = abs(headingStart - headingEnd)
                         
-                        let areaMinMax: [Double] = getSearchAreaMinMax(xyMinMax: xyMinMax, heading: [headingStart, headingEnd], recentScc: recentScc, searchType: isStraight, mode: mode)
+                        let areaMinMax: [Double] = getSearchAreaMinMax(xyMinMax: xyMinMax, heading: [headingStart, headingEnd], recentScc: recentScc, searchType: isStraight, lengthCondition: USER_TRAJECTORY_LENGTH)
                         let searchArea = getSearchCoordinates(areaMinMax: areaMinMax, interval: 1.0)
                         
                         let searchHeadings: [Double] = [compensateHeading(heading: headingEnd)]
@@ -3018,7 +2980,7 @@ public class ServiceManager: Observation {
                         let headingEnd = compensateHeading(heading: headingFromHead[0]-180)
                         let diffHeading = abs(90 - abs(headingStart - headingEnd))
                         
-                        let areaMinMax: [Double] = getSearchAreaMinMax(xyMinMax: xyMinMax, heading: [headingStart, headingEnd], recentScc: recentScc, searchType: isStraight, mode: mode)
+                        let areaMinMax: [Double] = getSearchAreaMinMax(xyMinMax: xyMinMax, heading: [headingStart, headingEnd], recentScc: recentScc, searchType: isStraight, lengthCondition: USER_TRAJECTORY_LENGTH)
                         let searchArea = getSearchCoordinates(areaMinMax: areaMinMax, interval: 1.0)
                         var searchHeadings: [Double] = []
                         
@@ -3076,7 +3038,7 @@ public class ServiceManager: Observation {
                         let diffX = xyMinMax[2] - xyMinMax[0]
                         let diffY = xyMinMax[3] - xyMinMax[1]
 
-                        let areaMinMax: [Double] = getSearchAreaMinMax(xyMinMax: xyMinMax, heading: [headingStart, headingEnd], recentScc: recentScc, searchType: isStraight, mode: mode)
+                        let areaMinMax: [Double] = getSearchAreaMinMax(xyMinMax: xyMinMax, heading: [headingStart, headingEnd], recentScc: recentScc, searchType: isStraight, lengthCondition: USER_TRAJECTORY_LENGTH)
                         let searchArea = getSearchCoordinates(areaMinMax: areaMinMax, interval: 1.0)
                         var searchHeadings: [Double] = []
                         
@@ -3143,7 +3105,7 @@ public class ServiceManager: Observation {
                         let headingEnd = compensateHeading(heading: headingFromHead[0]-180)
                         let diffHeading = abs(headingStart - headingEnd)
                         
-                        let areaMinMax: [Double] = getSearchAreaMinMax(xyMinMax: xyMinMax, heading: [headingStart, headingEnd], recentScc: recentScc, searchType: isStraight, mode: mode)
+                        let areaMinMax: [Double] = getSearchAreaMinMax(xyMinMax: xyMinMax, heading: [headingStart, headingEnd], recentScc: recentScc, searchType: isStraight, lengthCondition: USER_TRAJECTORY_LENGTH)
                         let searchArea = getSearchCoordinates(areaMinMax: areaMinMax, interval: 1.0)
                         let searchHeadings: [Double] = [compensateHeading(heading: headingEnd)]
                         
@@ -3194,8 +3156,6 @@ public class ServiceManager: Observation {
             }
         }
         
-//        resultRange = convertToValidSearchRange(inputRange: resultRange, pathPointMinMax: self.PathPointMinMax)
-        
         return (resultRange, resultDirection, tailIndex, searchType, resultRatio)
     }
     
@@ -3217,140 +3177,6 @@ public class ServiceManager: Observation {
         } else {
             return true
         }
-    }
-    
-    func getSearchAreaMinMax(xyMinMax: [Double], heading: [Double], recentScc: Double, searchType: Int, mode: String) -> [Double] {
-        var areaMinMax: [Double] = []
-        
-        var xMin = xyMinMax[0]
-        var yMin = xyMinMax[1]
-        var xMax = xyMinMax[2]
-        var yMax = xyMinMax[3]
-        
-        var lengthCondition = USER_TRAJECTORY_LENGTH
-        if (mode == "pdr") {
-            lengthCondition = USER_TRAJECTORY_DIAGONAL
-        }
-        let SEARCH_LENGTH: Double = lengthCondition*0.4
-        
-        let headingStart = heading[0]
-        let headingEnd = heading[1]
-
-        let startCos = cos(headingStart*D2R)
-        let startSin = sin(headingStart*D2R)
-
-        let endCos = cos(headingEnd*D2R)
-        let endSin = sin(headingEnd*D2R)
-        
-        if (searchType == 3) {
-            // Tail Straight
-            if (startCos > 0) {
-                xMin = xMin - SEARCH_LENGTH*startCos
-                xMax = xMax + SEARCH_LENGTH*startCos
-            } else {
-                xMin = xMin + SEARCH_LENGTH*startCos
-                xMax = xMax - SEARCH_LENGTH*startCos
-            }
-
-            if (startSin > 0) {
-                yMin = yMin - SEARCH_LENGTH*startSin
-                yMax = yMax + SEARCH_LENGTH*startSin
-            } else {
-                yMin = yMin + SEARCH_LENGTH*startSin
-                yMax = yMax - SEARCH_LENGTH*startSin
-            }
-
-            if (endCos > 0) {
-                xMin = xMin - 1.2*SEARCH_LENGTH*endCos
-                xMax = xMax + 1.2*SEARCH_LENGTH*endCos
-            } else {
-                xMin = xMin + 1.2*SEARCH_LENGTH*endCos
-                xMax = xMax - 1.2*SEARCH_LENGTH*endCos
-            }
-
-            if (endSin > 0) {
-                yMin = yMin - 1.2*SEARCH_LENGTH*endSin
-                yMax = yMax + 1.2*SEARCH_LENGTH*endSin
-            } else {
-                yMin = yMin + 1.2*SEARCH_LENGTH*endSin
-                yMax = yMax - 1.2*SEARCH_LENGTH*endSin
-            }
-        } else {
-            // All & Head Straight
-            if (startCos > 0) {
-                xMin = xMin - 1.2*SEARCH_LENGTH*startCos
-                xMax = xMax + 1.2*SEARCH_LENGTH*startCos
-            } else {
-                xMin = xMin + 1.2*SEARCH_LENGTH*startCos
-                xMax = xMax - 1.2*SEARCH_LENGTH*startCos
-            }
-
-            if (startSin > 0) {
-                yMin = yMin - 1.2*SEARCH_LENGTH*startSin
-                yMax = yMax + 1.2*SEARCH_LENGTH*startSin
-            } else {
-                yMin = yMin + 1.2*SEARCH_LENGTH*startSin
-                yMax = yMax - 1.2*SEARCH_LENGTH*startSin
-            }
-
-            if (endCos > 0) {
-                xMin = xMin - SEARCH_LENGTH*endCos
-                xMax = xMax + SEARCH_LENGTH*endCos
-            } else {
-                xMin = xMin + SEARCH_LENGTH*endCos
-                xMax = xMax - SEARCH_LENGTH*endCos
-            }
-
-            if (endSin > 0) {
-                yMin = yMin - SEARCH_LENGTH*endSin
-                yMax = yMax + SEARCH_LENGTH*endSin
-            } else {
-                yMin = yMin + SEARCH_LENGTH*endSin
-                yMax = yMax - SEARCH_LENGTH*endSin
-            }
-        }
-        
-        // 직선인 경우
-        if (abs(xMin - xMax) < 5.0) {
-            xMin = xMin - lengthCondition*0.05
-            xMax = xMax + lengthCondition*0.05
-        }
-
-        if (abs(yMin - yMax) < 5.0) {
-            yMin = yMin - lengthCondition*0.05
-            yMax = yMax + lengthCondition*0.05
-        }
-        
-        // U-Turn인 경우
-        let diffHeading = compensateHeading(heading: abs(headingStart - headingEnd))
-        let diffX = abs(xMax - xMin)
-        let diffY = abs(yMax - yMin)
-        let diffXy = abs(diffX - diffY)*0.2
-        
-        if (diffHeading > 150) {
-            if (diffX < diffY) {
-                xMin = xMin - diffXy
-                xMax = xMax + diffXy
-            } else {
-                yMin = yMin - diffXy
-                yMax = yMax + diffXy
-            }
-        } else {
-            // Check ㄹ Trajectory
-            if (diffHeading < 30 && searchType != 1) {
-                if (diffX < diffY) {
-                    xMin = xMin - diffXy
-                    xMax = xMax + diffXy
-                } else {
-                    yMin = yMin - diffXy
-                    yMax = yMax + diffXy
-                }
-            }
-        }
-
-        areaMinMax = [xMin, yMin, xMax, yMax]
-        
-        return areaMinMax
     }
     
     private func processPhase2(currentTime: Int, localTime: String, userTrajectory: [TrajectoryInfo], searchInfo: ([Int], [Int], Int, Int, Double)) {
@@ -3402,7 +3228,8 @@ public class ServiceManager: Observation {
                             self.scCompensationBadCount += 1
                         } else {
                             if (result.scc > 0.7) {
-                                self.scCompensation = result.sc_compensation
+                                let digit: Double = pow(10, 4)
+                                self.scCompensation = round((result.sc_compensation*digit)/digit)
                             }
                             self.scCompensationBadCount = 0
                         }
@@ -3430,16 +3257,16 @@ public class ServiceManager: Observation {
                         }
                         
                         self.pastSearchDirection = result.search_direction
-                        var resultHeading = compensateHeading(heading: result.absolute_heading)
-                        var resultCorrected = (true, [result.x, result.y, resultHeading])
+                        let resultHeading = compensateHeading(heading: result.absolute_heading)
+                        var resultCorrected = (true, [result.x, result.y, resultHeading, 1.0])
                         if (self.runMode == "pdr") {
-                            let pathMatchingResult = self.pathMatching(building: result.building_name, level: result.level_name, x: result.x, y: result.y, heading: resultHeading, tuXY: [0,0], isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: false, pathType: 0)
+                            let pathMatchingResult = pmCalculator.pathMatching(building: result.building_name, level: result.level_name, x: result.x, y: result.y, heading: resultHeading, isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: false, pathType: 0, range: SQUARE_RANGE)
                             resultCorrected.0 = pathMatchingResult.isSuccess
-                            resultCorrected.1 = pathMatchingResult.xyh
+                            resultCorrected.1 = pathMatchingResult.xyhs
                         } else {
-                            let pathMatchingResult = self.pathMatching(building: result.building_name, level: result.level_name, x: result.x, y: result.y, heading: resultHeading, tuXY: [0,0], isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: true, pathType: 1)
+                            let pathMatchingResult = pmCalculator.pathMatching(building: result.building_name, level: result.level_name, x: result.x, y: result.y, heading: resultHeading, isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: true, pathType: 1, range: SQUARE_RANGE)
                             resultCorrected.0 = pathMatchingResult.isSuccess
-                            resultCorrected.1 = pathMatchingResult.xyh
+                            resultCorrected.1 = pathMatchingResult.xyhs
                             
                             let isResultStraight = isResultHeadingStraight(drBuffer: self.unitDrBuffer, result: result)
                             if (!isResultStraight) {
@@ -3447,8 +3274,6 @@ public class ServiceManager: Observation {
                             }
                         }
                         resultCorrected.1[2] = compensateHeading(heading: resultCorrected.1[2])
-                        var resultHeadingMm = resultCorrected.1[2]
-                        
                         if (result.phase == 2 && result.scc < 0.25) {
                             self.isNeedTrajInit = true
                             self.phase = 1
@@ -3501,8 +3326,8 @@ public class ServiceManager: Observation {
                                     let propagationValues: [Double] = propagationResult.1
                                     if (propagationResult.0) {
                                         var propagatedResult: [Double] = [resultCorrected.1[0]+propagationValues[0] , resultCorrected.1[1]+propagationValues[1], resultCorrected.1[2]+propagationValues[2]]
-                                        let pathMatchingResult = self.pathMatching(building: result.building_name, level: result.level_name, x: propagatedResult[0], y: propagatedResult[1], heading: propagatedResult[2], tuXY: [0,0], isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: true, pathType: 1)
-                                        propagatedResult = pathMatchingResult.xyh
+                                        let pathMatchingResult = pmCalculator.pathMatching(building: result.building_name, level: result.level_name, x: propagatedResult[0], y: propagatedResult[1], heading: propagatedResult[2], isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: true, pathType: 1, range: SQUARE_RANGE)
+                                        propagatedResult = pathMatchingResult.xyhs
                                         propagatedResult[2] = compensateHeading(heading: propagatedResult[2])
                                         
                                         self.timeUpdatePosition.x = propagatedResult[0]
@@ -3557,11 +3382,11 @@ public class ServiceManager: Observation {
                                         if (propagationResult.0) {
                                             var propagatedResult: [Double] = [resultCorrected.1[0]+propagationValues[0] , resultCorrected.1[1]+propagationValues[1], resultCorrected.1[2]+propagationValues[2]]
                                             if (self.runMode == "pdr") {
-                                                let pathMatchingResult = self.pathMatching(building: result.building_name, level: result.level_name, x: propagatedResult[0], y: propagatedResult[1], heading: propagatedResult[2], tuXY: [0,0], isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: true, pathType: 0)
-                                                propagatedResult = pathMatchingResult.xyh
+                                                let pathMatchingResult = pmCalculator.pathMatching(building: result.building_name, level: result.level_name, x: propagatedResult[0], y: propagatedResult[1], heading: propagatedResult[2], isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: true, pathType: 0, range: SQUARE_RANGE)
+                                                propagatedResult = pathMatchingResult.xyhs
                                             } else {
-                                                let pathMatchingResult = self.pathMatching(building: result.building_name, level: result.level_name, x: propagatedResult[0], y: propagatedResult[1], heading: propagatedResult[2], tuXY: [0,0], isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: true, pathType: 1)
-                                                propagatedResult = pathMatchingResult.xyh
+                                                let pathMatchingResult = pmCalculator.pathMatching(building: result.building_name, level: result.level_name, x: propagatedResult[0], y: propagatedResult[1], heading: propagatedResult[2], isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: true, pathType: 1, range: SQUARE_RANGE)
+                                                propagatedResult = pathMatchingResult.xyhs
                                             }
                                             propagatedResult[2] = compensateHeading(heading: propagatedResult[2])
                                             
@@ -3720,7 +3545,8 @@ public class ServiceManager: Observation {
                                 self.scCompensationBadCount += 1
                             } else {
                                 if (result.scc > 0.7) {
-                                    self.scCompensation = result.sc_compensation
+                                    let digit: Double = pow(10, 4)
+                                    self.scCompensation = round((result.sc_compensation*digit)/digit)
                                 }
                                 self.scCompensationBadCount = 0
                             }
@@ -3785,16 +3611,16 @@ public class ServiceManager: Observation {
                         }
                         
                         self.pastSearchDirection = result.search_direction
-                        var resultHeading = compensateHeading(heading: result.absolute_heading)
-                        var resultCorrected = (true, [result.x, result.y, resultHeading])
+                        let resultHeading = compensateHeading(heading: result.absolute_heading)
+                        var resultCorrected = (true, [result.x, result.y, resultHeading, 1.0])
                         if (self.runMode == "pdr") {
-                            let pathMatchingResult = self.pathMatching(building: result.building_name, level: result.level_name, x: result.x, y: result.y, heading: result.absolute_heading, tuXY: [0,0], isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: false, pathType: 0)
+                            let pathMatchingResult = pmCalculator.pathMatching(building: result.building_name, level: result.level_name, x: result.x, y: result.y, heading: result.absolute_heading, isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: false, pathType: 0, range: SQUARE_RANGE)
                             resultCorrected.0 = pathMatchingResult.isSuccess
-                            resultCorrected.1 = pathMatchingResult.xyh
+                            resultCorrected.1 = pathMatchingResult.xyhs
                         } else {
-                            let pathMatchingResult = self.pathMatching(building: result.building_name, level: result.level_name, x: result.x, y: result.y, heading: result.absolute_heading, tuXY: [0,0], isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: true, pathType: 1)
+                            let pathMatchingResult = pmCalculator.pathMatching(building: result.building_name, level: result.level_name, x: result.x, y: result.y, heading: result.absolute_heading, isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: true, pathType: 1, range: SQUARE_RANGE)
                             resultCorrected.0 = pathMatchingResult.isSuccess
-                            resultCorrected.1 = pathMatchingResult.xyh
+                            resultCorrected.1 = pathMatchingResult.xyhs
                             
                             let isResultStraight = isResultHeadingStraight(drBuffer: self.unitDrBuffer, result: result)
                             if (!isResultStraight) {
@@ -3802,8 +3628,6 @@ public class ServiceManager: Observation {
                             }
                         }
                         resultCorrected.1[2] = compensateHeading(heading: resultCorrected.1[2])
-                        var resultHeadingMm = resultCorrected.1[2]
-                        
                         self.serverResult[0] = resultCorrected.1[0]
                         self.serverResult[1] = resultCorrected.1[1]
                         self.serverResult[2] = resultCorrected.1[2]
@@ -3838,11 +3662,11 @@ public class ServiceManager: Observation {
                                 if (propagationResult.0) {
                                     var propagatedResult: [Double] = [resultCorrected.1[0]+propagationValues[0] , resultCorrected.1[1]+propagationValues[1], resultCorrected.1[2]+propagationValues[2]]
                                     if (self.runMode == "pdr") {
-                                        let pathMatchingResult = self.pathMatching(building: result.building_name, level: result.level_name, x: propagatedResult[0], y: propagatedResult[1], heading: propagatedResult[2], tuXY: [0,0], isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: true, pathType: 0)
-                                        propagatedResult = pathMatchingResult.xyh
+                                        let pathMatchingResult = pmCalculator.pathMatching(building: result.building_name, level: result.level_name, x: propagatedResult[0], y: propagatedResult[1], heading: propagatedResult[2], isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: true, pathType: 0, range: SQUARE_RANGE)
+                                        propagatedResult = pathMatchingResult.xyhs
                                     } else {
-                                        let pathMatchingResult = self.pathMatching(building: result.building_name, level: result.level_name, x: propagatedResult[0], y: propagatedResult[1], heading: propagatedResult[2], tuXY: [0,0], isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: true, pathType: 1)
-                                        propagatedResult = pathMatchingResult.xyh
+                                        let pathMatchingResult = pmCalculator.pathMatching(building: result.building_name, level: result.level_name, x: propagatedResult[0], y: propagatedResult[1], heading: propagatedResult[2], isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: true, pathType: 1, range: SQUARE_RANGE)
+                                        propagatedResult = pathMatchingResult.xyhs
                                     }
                                     propagatedResult[2] = compensateHeading(heading: propagatedResult[2])
                                     
@@ -3936,8 +3760,8 @@ public class ServiceManager: Observation {
                             let propagationResult = propagateUsingUvd(drBuffer: self.unitDrBuffer, result: result)
                             let propagationValues: [Double] = propagationResult.1
                             var propagatedResult: [Double] = [resultCorrected.1[0]+propagationValues[0] , resultCorrected.1[1]+propagationValues[1], resultCorrected.1[2]+propagationValues[2]]
-                            let pathMatchingResult = self.pathMatching(building: result.building_name, level: result.level_name, x: propagatedResult[0], y: propagatedResult[1], heading: propagatedResult[2], tuXY: [0,0], isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: true, pathType: 1)
-                            propagatedResult = pathMatchingResult.xyh
+                            let pathMatchingResult = pmCalculator.pathMatching(building: result.building_name, level: result.level_name, x: propagatedResult[0], y: propagatedResult[1], heading: propagatedResult[2], isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: true, pathType: 1, range: SQUARE_RANGE)
+                            propagatedResult = pathMatchingResult.xyhs
                             propagatedResult[2] = compensateHeading(heading: propagatedResult[2])
                             
                             
@@ -4140,16 +3964,16 @@ public class ServiceManager: Observation {
                                     // Measurement Update 하기전에 현재 Time Update 위치를 고려
                                     var resultForMu = result
                                     resultForMu.absolute_heading = compensateHeading(heading: resultForMu.absolute_heading)
-                                    var resultHeading =  resultForMu.absolute_heading
-                                    var resultCorrected = (true, [resultForMu.x, resultForMu.y, resultForMu.absolute_heading])
+                                    let resultHeading =  resultForMu.absolute_heading
+                                    var resultCorrected = (true, [resultForMu.x, resultForMu.y, resultForMu.absolute_heading, 1.0])
                                     if (self.runMode == "pdr") {
-                                        let pathMatchingResult = self.pathMatching(building: resultForMu.building_name, level: resultForMu.level_name, x: resultForMu.x, y: resultForMu.y, heading: resultForMu.absolute_heading, tuXY: [0,0], isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: false, pathType: 0)
+                                        let pathMatchingResult = pmCalculator.pathMatching(building: resultForMu.building_name, level: resultForMu.level_name, x: resultForMu.x, y: resultForMu.y, heading: resultForMu.absolute_heading, isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: false, pathType: 0, range: SQUARE_RANGE)
                                         resultCorrected.0 = pathMatchingResult.isSuccess
-                                        resultCorrected.1 = pathMatchingResult.xyh
+                                        resultCorrected.1 = pathMatchingResult.xyhs
                                     } else {
-                                        let pathMatchingResult = self.pathMatching(building: resultForMu.building_name, level: resultForMu.level_name, x: resultForMu.x, y: resultForMu.y, heading: resultForMu.absolute_heading, tuXY: [0,0], isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: true, pathType: 1)
+                                        let pathMatchingResult = pmCalculator.pathMatching(building: resultForMu.building_name, level: resultForMu.level_name, x: resultForMu.x, y: resultForMu.y, heading: resultForMu.absolute_heading, isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: true, pathType: 1, range: SQUARE_RANGE)
                                         resultCorrected.0 = pathMatchingResult.isSuccess
-                                        resultCorrected.1 = pathMatchingResult.xyh
+                                        resultCorrected.1 = pathMatchingResult.xyhs
                                         
                                         if (inputTrajType == 0) {
                                             let isResultStraight = isResultHeadingStraight(drBuffer: self.unitDrBuffer, result: result)
@@ -4159,8 +3983,6 @@ public class ServiceManager: Observation {
                                         }
                                     }
                                     resultCorrected.1[2] = compensateHeading(heading: resultCorrected.1[2])
-                                    var resultHeadingMm = resultCorrected.1[2]
-                                            
                                     self.serverResult[0] = resultCorrected.1[0]
                                     self.serverResult[1] = resultCorrected.1[1]
                                     self.serverResult[2] = resultCorrected.1[2]
@@ -4576,36 +4398,6 @@ public class ServiceManager: Observation {
         return levelArrayToReturn
     }
     
-    func checkInEntranceMatchingArea(x: Double, y: Double, building: String, level: String) -> (Bool, [Double]) {
-        var area = [Double]()
-        
-        let buildingName = building
-        let levelName = removeLevelDirectionString(levelName: level)
-        
-        let key = "\(buildingName)_\(levelName)"
-        guard let entranceMatchingArea: [[Double]] = EntranceMatchingArea[key] else {
-            return (false, area)
-        }
-        
-        for i in 0..<entranceMatchingArea.count {
-            if (!entranceMatchingArea[i].isEmpty) {
-                let xMin = entranceMatchingArea[i][0]
-                let yMin = entranceMatchingArea[i][1]
-                let xMax = entranceMatchingArea[i][2]
-                let yMax = entranceMatchingArea[i][3]
-                
-                if (x >= xMin && x <= xMax) {
-                    if (y >= yMin && y <= yMax) {
-                        area = entranceMatchingArea[i]
-                        return (true, area)
-                    }
-                }
-            }
-        }
-        
-        return (false, area)
-    }
-    
     func postParam(sector_id: Int, normailzationScale: Double) {
         let localTime = getLocalTimeString()
         
@@ -4673,74 +4465,6 @@ public class ServiceManager: Observation {
         }
     }
     
-    func fromServerToResult(fromServer: FineLocationTrackingFromServer, velocity: Double) -> FineLocationTrackingResult {
-        var result = FineLocationTrackingResult()
-        
-        result.mobile_time = fromServer.mobile_time
-        result.building_name = fromServer.building_name
-        result.level_name = fromServer.level_name
-        result.scc = fromServer.scc
-        result.x = fromServer.x
-        result.y = fromServer.y
-        result.absolute_heading = fromServer.absolute_heading
-        result.phase = fromServer.phase
-        result.calculated_time = fromServer.calculated_time
-        result.index = fromServer.index
-        result.velocity = velocity
-        
-        return result
-    }
-    
-    func jsonToInfoResult(json: String) -> InfoResult {
-        let result = InfoResult.init()
-        let decoder = JSONDecoder()
-        
-        let jsonString = json
-        
-        if let data = jsonString.data(using: .utf8), let decoded = try? decoder.decode(InfoResult.self, from: data) {
-            return decoded
-        }
-        
-        return result
-    }
-    
-    private func parseRoad(data: String) -> ([Int], [[Double]], [Double], [String] ) {
-        var roadType = [Int]()
-        var road = [[Double]]()
-        var roadScale = [Double]()
-        var roadHeading = [String]()
-        
-        var roadX = [Double]()
-        var roadY = [Double]()
-        
-        let roadString = data.components(separatedBy: .newlines)
-        for i in 0..<roadString.count {
-            if (roadString[i] != "") {
-                let lineData = roadString[i].components(separatedBy: ",")
-                
-                roadType.append(Int(Double(lineData[0])!))
-                roadX.append(Double(lineData[1])!)
-                roadY.append(Double(lineData[2])!)
-                roadScale.append(Double(lineData[3])!)
-                
-                var headingArray: String = ""
-                if (lineData.count > 4) {
-                    for j in 4..<lineData.count {
-                        headingArray.append(lineData[j])
-                        if (lineData[j] != "") {
-                            headingArray.append(",")
-                        }
-                    }
-                }
-                roadHeading.append(headingArray)
-            }
-        }
-        road = [roadX, roadY]
-        self.PathPointMinMax = [roadX.min() ?? 0, roadY.min() ?? 0, roadX.max() ?? 0, roadY.max() ?? 0]
-        
-        return (roadType, road, roadScale, roadHeading)
-    }
-    
     private func parseEntrance(data: String) -> ([String], [[Double]]) {
         var entracneLevelArray = [String]()
         var entranceArray = [[Double]]()
@@ -4767,10 +4491,10 @@ public class ServiceManager: Observation {
         let buildingName = result.building_name
         let levelName = removeLevelDirectionString(levelName: result.level_name)
         
-        let resultPm = self.pathMatching(building: buildingName, level: levelName, x: result.x, y: result.y, heading: result.absolute_heading, tuXY: [0, 0], isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: false, pathType: 1)
+        let resultPm = pmCalculator.pathMatching(building: buildingName, level: levelName, x: result.x, y: result.y, heading: result.absolute_heading, isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: false, pathType: 1, range: SQUARE_RANGE)
         
-        let coordX = resultPm.xyh[0]
-        let coordY = resultPm.xyh[1]
+        let coordX = resultPm.xyhs[0]
+        let coordY = resultPm.xyhs[1]
         
         var resultCopy = result
         resultCopy.x = coordX
@@ -4918,7 +4642,7 @@ public class ServiceManager: Observation {
     }
     
     private func determineIsOutdoor(lastResult: FineLocationTrackingResult, currentTime: Double, inFailCondition: Bool) -> Bool {
-        let isInEntranceMatchingArea = self.checkInEntranceMatchingArea(x: lastResult.x, y: lastResult.y, building: lastResult.building_name, level: lastResult.level_name)
+        let isInEntranceMatchingArea = pmCalculator.checkInEntranceMatchingArea(x: lastResult.x, y: lastResult.y, building: lastResult.building_name, level: lastResult.level_name)
         
         let diffEntranceWardTime = currentTime - self.lastScannedEntranceOuterWardTime
         if (lastResult.building_name != "" && lastResult.level_name == "B0") {
@@ -4944,798 +4668,7 @@ public class ServiceManager: Observation {
         return false
     }
     
-    public func pathMatching(building: String, level: String, x: Double, y: Double, heading: Double, tuXY: [Double], isPast: Bool, HEADING_RANGE: Double, isUseHeading: Bool, pathType: Int) -> (isSuccess: Bool, xyh: [Double]) {
-        var isSuccess: Bool = false
-        var xyh: [Double] = [x, y, heading]
-        let levelCopy: String = removeLevelDirectionString(levelName: level)
-        let key: String = "\(building)_\(levelCopy)"
-        if (isPast) {
-            isSuccess = true
-            return (isSuccess, xyh)
-        }
-        
-        if (!(building.isEmpty) && !(level.isEmpty)) {
-            guard let mainType: [Int] = self.PathType[key] else {
-                return (isSuccess, xyh)
-            }
-            guard let mainRoad: [[Double]] = self.PathPoint[key] else {
-                return (isSuccess, xyh)
-            }
-            
-            guard let mainMagScale: [Double] = self.PathMagScale[key] else {
-                return (isSuccess, xyh)
-            }
-            
-            guard let mainHeading: [String] = self.PathHeading[key] else {
-                return (isSuccess, xyh)
-            }
-            
-            let pathhMatchingArea = checkInEntranceMatchingArea(x: x, y: y, building: building, level: levelCopy)
-            
-            var idshArray = [[Double]]()
-            var idshArrayWhenFail = [[Double]]()
-            var pathArray = [[Double]]()
-            if (!mainRoad.isEmpty) {
-                let roadX = mainRoad[0]
-                let roadY = mainRoad[1]
-                
-                var xMin = x - SQUARE_RANGE
-                var xMax = x + SQUARE_RANGE
-                var yMin = y - SQUARE_RANGE
-                var yMax = y + SQUARE_RANGE
-                if (pathhMatchingArea.0) {
-                    xMin = pathhMatchingArea.1[0]
-                    yMin = pathhMatchingArea.1[1]
-                    xMax = pathhMatchingArea.1[2]
-                    yMax = pathhMatchingArea.1[3]
-                }
-                
-                for i in 0..<roadX.count {
-                    let xPath = roadX[i]
-                    let yPath = roadY[i]
-                    
-                    let pathTypeLoaded = mainType[i]
-                    if (pathType == 1) {
-                        if (pathType != pathTypeLoaded) {
-                            continue
-                        }
-                    }
-                    // XY 범위 안에 있는 값 중에 검사
-                    if (xPath >= xMin && xPath <= xMax) {
-                        if (yPath >= yMin && yPath <= yMax) {
-                            let index = Double(i)
-                            let distance = sqrt(pow(x-xPath, 2) + pow(y-yPath, 2))
-                            
-                            let magScale = mainMagScale[i]
-                            var idsh: [Double] = [index, distance, magScale, heading]
-                            var path: [Double] = [xPath, yPath, 0, 0]
-                            
-                            idshArrayWhenFail.append(idsh)
-                            
-                            // Heading 사용
-                            if (isUseHeading) {
-                                let headingArray = mainHeading[i]
-                                var isValidIdh: Bool = true
-                                if (!headingArray.isEmpty) {
-                                    let headingData = headingArray.components(separatedBy: ",")
-                                    var diffHeading = [Double]()
-                                    for j in 0..<headingData.count {
-                                        if(!headingData[j].isEmpty) {
-                                            let mapHeading = Double(headingData[j])!
-                                            if (heading > 270 && (mapHeading >= 0 && mapHeading < 90)) {
-                                                diffHeading.append(abs(heading - (mapHeading+360)))
-                                            } else if (mapHeading > 270 && (heading >= 0 && heading < 90)) {
-                                                diffHeading.append(abs(mapHeading - (heading+360)))
-                                            } else {
-                                                diffHeading.append(abs(heading - mapHeading))
-                                            }
-                                        }
-                                    }
-                                    
-                                    if (!diffHeading.isEmpty) {
-                                        let idxHeading = diffHeading.firstIndex(of: diffHeading.min()!)
-                                        let minHeading = Double(headingData[idxHeading!])!
-                                        idsh[3] = minHeading
-                                        if (isUseHeading) {
-                                            if (heading > 270 && (minHeading >= 0 && minHeading < 90)) {
-                                                if (abs(minHeading+360-heading) >= HEADING_RANGE) {
-                                                    isValidIdh = false
-                                                }
-                                            } else if (minHeading > 270 && (heading >= 0 && heading < 90)) {
-                                                if (abs(heading+360-minHeading) >= HEADING_RANGE) {
-                                                    isValidIdh = false
-                                                }
-                                            } else {
-                                                if (abs(heading-minHeading) >= HEADING_RANGE) {
-                                                    isValidIdh = false
-                                                }
-                                            }
-                                        }
-                                        path[2] = minHeading
-                                        path[3] = 1
-                                    }
-                                }
-                                
-                                if (isValidIdh) {
-                                    idshArray.append(idsh)
-                                    pathArray.append(path)
-                                }
-                                
-                                if (!idshArray.isEmpty) {
-                                    let sortedIdsh = idshArray.sorted(by: {$0[1] < $1[1] })
-                                    var index: Int = 0
-                                    var correctedHeading: Double = heading
-                                    var correctedScale = 1.0
-                                    
-                                    if (!sortedIdsh.isEmpty) {
-                                        let minData: [Double] = sortedIdsh[0]
-                                        index = Int(minData[0])
-                                        if (isUseHeading) {
-                                            correctedScale = minData[2]
-                                            correctedHeading = minData[3]
-                                        } else {
-                                            correctedHeading = heading
-                                        }
-                                    }
-                                    
-                                    isSuccess = true
-                                    
-                                    if (correctedScale < 0.7) {
-                                        correctedScale = 0.7
-                                    }
-                                    
-                                    unitDRGenerator.setVelocityScaleFactor(scaleFactor: correctedScale)
-                                    xyh = [roadX[index], roadY[index], correctedHeading]
-                                } else {
-                                    let sortedIdsh = idshArrayWhenFail.sorted(by: {$0[1] < $1[1] })
-                                    var index: Int = 0
-                                    var correctedScale = 1.0
-                                    
-                                    if (!sortedIdsh.isEmpty) {
-                                        let minData: [Double] = sortedIdsh[0]
-                                        index = Int(minData[0])
-                                        correctedScale = minData[2]
-                                    }
-                                    
-                                    isSuccess = false
-                                    
-                                    if (correctedScale < 0.7) {
-                                        correctedScale = 0.7
-                                    }
-                                    
-                                    unitDRGenerator.setVelocityScaleFactor(scaleFactor: correctedScale)
-                                    xyh = [roadX[index], roadY[index], heading]
-                                }
-                            } else {
-                                // Heading 미사용
-                                idshArray.append(idsh)
-                                pathArray.append(path)
-                                if (!idshArray.isEmpty) {
-                                    isSuccess = true
-                                    
-                                    let sortedIdsh = idshArray.sorted(by: {$0[1] < $1[1] })
-                                    var index: Int = 0
-                                    if (!sortedIdsh.isEmpty) {
-                                        let minData: [Double] = sortedIdsh[0]
-                                        index = Int(minData[0])
-                                        xyh = [roadX[index], roadY[index], heading]
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-        return (isSuccess, xyh)
-    }
     
-    func pathTrajectoryMatching(building: String, level: String, x: Double, y: Double, heading: Double, pastResult: FineLocationTrackingResult, drBuffer: [UnitDRInfo], HEADING_RANGE: Double, pathType: Int) -> (isSuccess: Bool, xyd: [Double], minTrajectory: [[Double]]) {
-        let pastX = pastResult.x
-        let pastY = pastResult.y
-        
-        var isSuccess: Bool = false
-        var xyd: [Double] = [x, y, 50]
-        var minTrajectory = [[Double]]()
-        
-        let levelCopy: String = removeLevelDirectionString(levelName: level)
-        let key: String = "\(building)_\(levelCopy)"
-        
-        if (!(building.isEmpty) && !(level.isEmpty)) {
-            guard let mainType: [Int] = self.PathType[key] else {
-                return (isSuccess, xyd, minTrajectory)
-            }
-            guard let mainRoad: [[Double]] = self.PathPoint[key] else {
-                return (isSuccess, xyd, minTrajectory)
-            }
-            
-            if (!mainRoad.isEmpty) {
-                let roadX = mainRoad[0]
-                let roadY = mainRoad[1]
-                
-                var xMin = x - SQUARE_RANGE
-                var xMax = x + SQUARE_RANGE
-                var yMin = y - SQUARE_RANGE
-                var yMax = y + SQUARE_RANGE
-                
-                var ppXydArray = [[Double]]()
-                var minDistanceCoord = [Double]()
-                
-                for i in 0..<roadX.count {
-                    let xPath = roadX[i]
-                    let yPath = roadY[i]
-                    
-                    let pathTypeLoaded = mainType[i]
-                    if (pathType == 1) {
-                        if (pathType != pathTypeLoaded) {
-                            continue
-                        }
-                    }
-                    
-                    // XY 범위 안에 있는 값 중에 검사
-                    if (xPath >= xMin && xPath <= xMax) {
-                        if (yPath >= yMin && yPath <= yMax) {
-                            var distanceSum: Double = 0
-                            
-                            let headingCompensation: Double = heading - drBuffer[drBuffer.count-1].heading
-                            var headingBuffer: [Double] = []
-                            for i in 0..<drBuffer.count {
-                                let compensatedHeading = compensateHeading(heading: drBuffer[i].heading + headingCompensation - 180)
-                                headingBuffer.append(compensatedHeading)
-                            }
-                            
-                            var xyFromHead: [Double] = [xPath, yPath]
-                            let firstXyd = calDistacneFromNearestPp(coord: xyFromHead, mainRoad: mainRoad, mainType: mainType, pathType: pathType)
-                            var xydArray: [[Double]] = [firstXyd]
-                            distanceSum += firstXyd[2]
-                            
-                            var trajectoryFromHead = [[Double]]()
-                            trajectoryFromHead.append(xyFromHead)
-                            for i in (1..<drBuffer.count).reversed() {
-                                let headAngle = headingBuffer[i]
-                                xyFromHead[0] = xyFromHead[0] + drBuffer[i].length*cos(headAngle*D2R)
-                                xyFromHead[1] = xyFromHead[1] + drBuffer[i].length*sin(headAngle*D2R)
-                                trajectoryFromHead.append(xyFromHead)
-                                
-                                
-                                let calculatedXyd = calDistacneFromNearestPp(coord: xyFromHead, mainRoad: mainRoad, mainType: mainType, pathType: pathType)
-                                xydArray.append(calculatedXyd)
-                                distanceSum += calculatedXyd[2]
-                            }
-                            let distWithPast = sqrt((pastX - xPath)*(pastX - xPath) + (pastY - yPath)*(pastY - yPath))
-                            ppXydArray.append([xPath, yPath, distanceSum, distWithPast])
-                            
-                            if (minDistanceCoord.isEmpty) {
-                                minDistanceCoord = [xPath, yPath, distanceSum, distWithPast]
-                                minTrajectory = trajectoryFromHead
-                            } else {
-                                let distanceCurrent = distanceSum
-                                let distancePast = minDistanceCoord[2]
-                                if (distanceCurrent < distancePast && distWithPast <= 3) {
-                                    minDistanceCoord = [xPath, yPath, distanceSum, distWithPast]
-                                    minTrajectory = trajectoryFromHead
-                                }
-                            }
-                        }
-                    }
-                    
-                    if (!minDistanceCoord.isEmpty) {
-                        if (minDistanceCoord[2] <= 15 && minDistanceCoord[3] <= 3) {
-                            isSuccess = true
-                        } else {
-                            isSuccess = false
-                        }
-                        xyd = minDistanceCoord
-                    }
-                }
-            }
-        }
-        return (isSuccess, xyd, minTrajectory)
-    }
-    
-    func extendedPathTrajectoryMatching(building: String, level: String, x: Double, y: Double, heading: Double, pastResult: FineLocationTrackingResult, drBuffer: [UnitDRInfo], HEADING_RANGE: Double, pathType: Int, mode: String) -> (isSuccess: Bool, xyd: [Double], minTrajectory: [[Double]], minTrajectoryOriginal: [[Double]]) {
-        let startTime: Double = getCurrentTimeInMillisecondsDouble()
-        
-        let pastX = pastResult.x
-        let pastY = pastResult.y
-        
-        var isSuccess: Bool = false
-        var xyd: [Double] = [x, y, 50]
-        var minTrajectory = [[Double]]()
-        var minTrajectoryOriginal = [[Double]]()
-        
-        let levelCopy: String = removeLevelDirectionString(levelName: level)
-        let key: String = "\(building)_\(levelCopy)"
-        
-        if (!(building.isEmpty) && !(level.isEmpty)) {
-            guard let mainType: [Int] = self.PathType[key] else {
-                return (isSuccess, xyd, minTrajectory, minTrajectoryOriginal)
-            }
-            guard let mainRoad: [[Double]] = self.PathPoint[key] else {
-                return (isSuccess, xyd, minTrajectory, minTrajectoryOriginal)
-            }
-            
-            if (!mainRoad.isEmpty) {
-                let roadX = mainRoad[0]
-                let roadY = mainRoad[1]
-                
-                var xMin = x - SQUARE_RANGE
-                var xMax = x + SQUARE_RANGE
-                var yMin = y - SQUARE_RANGE
-                var yMax = y + SQUARE_RANGE
-                
-                var ppXydArray = [[Double]]()
-                var minDistanceCoord = [Double]()
-                
-                for i in 0..<roadX.count {
-                    let xPath = roadX[i]
-                    let yPath = roadY[i]
-                    
-                    let pathTypeLoaded = mainType[i]
-                    if (pathType == 1) {
-                        if (pathType != pathTypeLoaded) {
-                            continue
-                        }
-                    }
-                    
-                    // XY 범위 안에 있는 값 중에 검사
-                    if (xPath >= xMin && xPath <= xMax) {
-                        if (yPath >= yMin && yPath <= yMax) {
-                            var passedPp = [[Double]]()
-                            var distanceSum: Double = 0
-                            
-                            let headingCompensation: Double = heading - drBuffer[drBuffer.count-1].heading
-                            var headingBuffer: [Double] = []
-                            for i in 0..<drBuffer.count {
-                                let compensatedHeading = compensateHeading(heading: drBuffer[i].heading + headingCompensation - 180)
-                                headingBuffer.append(compensatedHeading)
-                            }
-                            
-                            var xyFromHead: [Double] = [xPath, yPath]
-                            var xyOriginal: [Double] = [xPath, yPath]
-                            let firstXyd = extendedCalDistacneFromNearestPp(coord: xyFromHead, passedPp: passedPp, mainRoad: mainRoad, mainType: mainType, pathType: pathType)
-                            passedPp.append(xyFromHead)
-                            
-                            var xydArray: [[Double]] = [firstXyd]
-                            distanceSum += firstXyd[2]
-                            
-                            var trajectoryFromHead = [[Double]]()
-                            var trajectoryOriginal = [[Double]]()
-                            trajectoryFromHead.append(xyFromHead)
-                            trajectoryOriginal.append(xyOriginal)
-                            for i in (1..<drBuffer.count).reversed() {
-                                let headAngle = headingBuffer[i]
-                                xyOriginal[0] = xyOriginal[0] + drBuffer[i].length*cos(headAngle*D2R)
-                                xyOriginal[1] = xyOriginal[1] + drBuffer[i].length*sin(headAngle*D2R)
-                                trajectoryOriginal.append(xyOriginal)
-                                if (mode == "pdr") {
-                                    if (i%2 == 0) {
-                                        let propagatedX = xyFromHead[0] + drBuffer[i].length*cos(headAngle*D2R)
-                                        let propagatedY = xyFromHead[1] + drBuffer[i].length*sin(headAngle*D2R)
-                                        let calculatedXyd = extendedCalDistacneFromNearestPp(coord: [propagatedX, propagatedY], passedPp: passedPp, mainRoad: mainRoad, mainType: mainType, pathType: pathType)
-                                        
-                                        xyFromHead[0] = calculatedXyd[0]
-                                        xyFromHead[1] = calculatedXyd[1]
-                                        xydArray.append(calculatedXyd)
-                                        distanceSum += calculatedXyd[2]
-                                        trajectoryFromHead.append(xyFromHead)
-                                        passedPp.append(xyFromHead)
-                                    } else {
-                                        let propagatedX = xyFromHead[0] + drBuffer[i].length*cos(headAngle*D2R)
-                                        let propagatedY = xyFromHead[1] + drBuffer[i].length*sin(headAngle*D2R)
-                                        let calculatedXyd = extendedCalDistacneFromNearestPp(coord: [propagatedX, propagatedY], passedPp: passedPp, mainRoad: mainRoad, mainType: mainType, pathType: pathType)
-                                        
-                                        xyFromHead[0] = propagatedX
-                                        xyFromHead[1] = propagatedY
-                                        xydArray.append(calculatedXyd)
-                                        distanceSum += calculatedXyd[2]
-                                        trajectoryFromHead.append(xyFromHead)
-                                        passedPp.append(xyFromHead)
-                                    }
-                                } else {
-                                    let propagatedX = xyFromHead[0] + drBuffer[i].length*cos(headAngle*D2R)
-                                    let propagatedY = xyFromHead[1] + drBuffer[i].length*sin(headAngle*D2R)
-                                    let calculatedXyd = extendedCalDistacneFromNearestPp(coord: [propagatedX, propagatedY], passedPp: passedPp, mainRoad: mainRoad, mainType: mainType, pathType: pathType)
-                                    
-                                    xyFromHead[0] = calculatedXyd[0]
-                                    xyFromHead[1] = calculatedXyd[1]
-                                    xydArray.append(calculatedXyd)
-                                    distanceSum += calculatedXyd[2]
-                                    trajectoryFromHead.append(xyFromHead)
-                                    passedPp.append(xyFromHead)
-                                }
-                            }
-                            
-                            let distWithPast = sqrt((pastX - xPath)*(pastX - xPath) + (pastY - yPath)*(pastY - yPath))
-                            ppXydArray.append([xPath, yPath, distanceSum, distWithPast])
-                            
-                            if (minDistanceCoord.isEmpty) {
-                                minDistanceCoord = [xPath, yPath, distanceSum, distWithPast]
-                                minTrajectory = trajectoryFromHead
-                                minTrajectoryOriginal = trajectoryOriginal
-                            } else {
-                                let distanceCurrent = distanceSum
-                                let distancePast = minDistanceCoord[2]
-                                if (distanceCurrent < distancePast && distWithPast <= 3) {
-                                    minDistanceCoord = [xPath, yPath, distanceSum, distWithPast]
-                                    minTrajectory = trajectoryFromHead
-                                    minTrajectoryOriginal = trajectoryOriginal
-                                }
-                            }
-                        }
-                    }
-                    
-//                    print(getLocalTimeString() + " , (PM) minDistanceCoord = \(minDistanceCoord)")
-                    if (!minDistanceCoord.isEmpty) {
-                        if (minDistanceCoord[2] <= 15 && minDistanceCoord[3] <= 3) {
-                            isSuccess = true
-                        } else {
-                            isSuccess = false
-                        }
-                        xyd = minDistanceCoord
-                    }
-                }
-            }
-        }
-        
-//        print(getLocalTimeString() + " , (PM) isSuccess = \(isSuccess) // minTrajectory = \(minTrajectory)")
-        let endTime: Double = getCurrentTimeInMillisecondsDouble()
-        let checkTime = (endTime-startTime)*1e-3
-//        print(getLocalTimeString() + " , (Jupiter) Path-Traj Matching : time = \(checkTime)")
-        return (isSuccess, xyd, minTrajectory, minTrajectoryOriginal)
-    }
-    
-    func calDistacneFromNearestPp(coord: [Double], mainRoad: [[Double]], mainType: [Int], pathType: Int) -> [Double] {
-        let x = coord[0]
-        let y = coord[1]
-        
-        var xyd: [Double] = [x, y, 50]
-        
-        var xydArray = [[Double]]()
-        
-        let roadX = mainRoad[0]
-        let roadY = mainRoad[1]
-        
-        var xMin = x - SQUARE_RANGE
-        var xMax = x + SQUARE_RANGE
-        var yMin = y - SQUARE_RANGE
-        var yMax = y + SQUARE_RANGE
-        
-        for i in 0..<roadX.count {
-            let xPath = roadX[i]
-            let yPath = roadY[i]
-            
-            let pathTypeLoaded = mainType[i]
-            if (pathType == 1) {
-                if (pathType != pathTypeLoaded) {
-                    continue
-                }
-            }
-            // XY 범위 안에 있는 값 중에 검사
-            if (xPath >= xMin && xPath <= xMax) {
-                if (yPath >= yMin && yPath <= yMax) {
-                    let distance = sqrt(pow(x-xPath, 2) + pow(y-yPath, 2))
-                    var xyd: [Double] = [xPath, yPath, distance]
-                    
-                    xydArray.append(xyd)
-                }
-            }
-        }
-        
-        if (!xydArray.isEmpty) {
-            let sortedXyd = xydArray.sorted(by: {$0[2] < $1[2] })
-            if (!sortedXyd.isEmpty) {
-                let minData: [Double] = sortedXyd[0]
-                xyd = minData
-            }
-        }
-        return xyd
-    }
-    
-    func extendedCalDistacneFromNearestPp(coord: [Double], passedPp: [[Double]], mainRoad: [[Double]], mainType: [Int], pathType: Int) -> [Double] {
-        let x = coord[0]
-        let y = coord[1]
-        
-        var xyd: [Double] = [x, y, 50]
-        
-        var xydArray = [[Double]]()
-        
-        let roadX = mainRoad[0]
-        let roadY = mainRoad[1]
-        
-        var xMin = x - SQUARE_RANGE
-        var xMax = x + SQUARE_RANGE
-        var yMin = y - SQUARE_RANGE
-        var yMax = y + SQUARE_RANGE
-        
-        for i in 0..<roadX.count {
-            let xPath = roadX[i]
-            let yPath = roadY[i]
-            
-            let pathTypeLoaded = mainType[i]
-            if (pathType == 1) {
-                if (pathType != pathTypeLoaded) {
-                    continue
-                }
-            }
-            
-            // XY 범위 안에 있는 값 중에 검사
-            if (!passedPp.isEmpty) {
-                let isContain: Bool = containsArray(passedPp, [xPath, yPath])
-                if (isContain) {
-                    continue
-                }
-            }
-            
-            if (xPath >= xMin && xPath <= xMax) {
-                if (yPath >= yMin && yPath <= yMax) {
-                    let distance = sqrt(pow(x-xPath, 2) + pow(y-yPath, 2))
-                    var xyd: [Double] = [xPath, yPath, distance]
-                    
-                    xydArray.append(xyd)
-                }
-            }
-        }
-        
-        if (!xydArray.isEmpty) {
-            let sortedXyd = xydArray.sorted(by: {$0[2] < $1[2] })
-            if (!sortedXyd.isEmpty) {
-                let minData: [Double] = sortedXyd[0]
-                xyd = minData
-            }
-        }
-        return xyd
-    }
-    
-    func calTrajectoryRatio(trajPm: [[Double]], trajOg: [[Double]]) -> Double {
-        var ratio = 1.0
-        
-        var lengthPm: Double = 0
-        var lengthOg: Double = 0
-        
-        for i in 1..<trajPm.count {
-            let pmDiffX = trajPm[i][0] - trajPm[i-1][0]
-            let pmDiffY = trajPm[i][1] - trajPm[i-1][1]
-            lengthPm += sqrt(pmDiffX*pmDiffX + pmDiffY*pmDiffY)
-            
-            let ogDiffX = trajOg[i][0] - trajOg[i-1][0]
-            let ogDiffY = trajOg[i][1] - trajOg[i-1][1]
-            lengthOg += sqrt(ogDiffX*ogDiffX + ogDiffY*ogDiffY)
-        }
-        
-        ratio = lengthPm/lengthOg
-        
-        return ratio
-    }
-    
-    func containsArray(_ array2D: [[Double]], _ targetArray: [Double]) -> Bool {
-        for array in array2D {
-            if array == targetArray {
-                return true
-            }
-        }
-        return false
-    }
-    
-    func getBestMatchedTrajectory(building: String, level: String, searchRange: [Double], tailHeading: Double, uvHeading: [Double], uvLength: [Double], pathType: Int, mode: String) -> MatchedTraj {
-        var isSuccess: Bool = false
-        var xyd: [Double] = [0, 0, 50]
-        var minTrajectory = [[Double]]()
-        var minTrajectoryOriginal = [[Double]]()
-        
-        var matchedTraj = MatchedTraj(isSuccess: isSuccess, xyd: xyd, minTrajectory: minTrajectory, minTrajectoryOriginal: minTrajectoryOriginal)
-        let levelCopy: String = removeLevelDirectionString(levelName: level)
-        let key: String = "\(building)_\(levelCopy)"
-        
-        if (!(building.isEmpty) && !(level.isEmpty)) {
-            guard let mainType: [Int] = self.PathType[key] else {
-                return matchedTraj
-            }
-            guard let mainRoad: [[Double]] = self.PathPoint[key] else {
-                return matchedTraj
-            }
-            
-            if (!mainRoad.isEmpty) {
-                let roadX = mainRoad[0]
-                let roadY = mainRoad[1]
-                
-                var xMin = searchRange[0]
-                var xMax = searchRange[2]
-                var yMin = searchRange[1]
-                var yMax = searchRange[3]
-                
-                var ppXydArray = [[Double]]()
-                var minDistanceCoord = [Double]()
-                
-                for i in 0..<roadX.count {
-                    let xPath = roadX[i]
-                    let yPath = roadY[i]
-                    
-                    let pathTypeLoaded = mainType[i]
-                    if (pathType == 1) {
-                        if (pathType != pathTypeLoaded) {
-                            continue
-                        }
-                    }
-                    
-                    // XY 범위 안에 있는 값 중에 검사
-                    if (xPath >= xMin && xPath <= xMax) {
-                        if (yPath >= yMin && yPath <= yMax) {
-                            var passedPp = [[Double]]()
-                            var distanceSum: Double = 0
-                            
-                            let headingCompensation: Double = tailHeading - uvHeading[0]
-                            var headingBuffer: [Double] = []
-                            for i in 0..<uvHeading.count {
-                                let compensatedHeading = compensateHeading(heading: uvHeading[i] + headingCompensation)
-                                headingBuffer.append(compensatedHeading)
-                            }
-                            
-                            var xyFromHead: [Double] = [xPath, yPath]
-                            var xyOriginal: [Double] = [xPath, yPath]
-                            let firstXyd = extendedCalDistacneFromNearestPp(coord: xyFromHead, passedPp: passedPp, mainRoad: mainRoad, mainType: mainType, pathType: pathType)
-                            passedPp.append(xyFromHead)
-                            
-                            var xydArray: [[Double]] = [firstXyd]
-                            distanceSum += firstXyd[2]
-                            
-                            var trajectoryFromHead = [[Double]]()
-                            var trajectoryOriginal = [[Double]]()
-                            trajectoryFromHead.append(xyFromHead)
-                            trajectoryOriginal.append(xyOriginal)
-                            for i in 1..<uvHeading.count{
-                                let headAngle = headingBuffer[i]
-                                xyOriginal[0] = xyOriginal[0] + uvLength[i]*cos(headAngle*D2R)
-                                xyOriginal[1] = xyOriginal[1] + uvLength[i]*sin(headAngle*D2R)
-                                trajectoryOriginal.append(xyOriginal)
-                                
-                                if (mode == "pdr") {
-                                    if (i%2 == 0) {
-                                        let propagatedX = xyFromHead[0] + uvLength[i]*cos(headAngle*D2R)
-                                        let propagatedY = xyFromHead[1] + uvLength[i]*sin(headAngle*D2R)
-                                        let calculatedXyd = extendedCalDistacneFromNearestPp(coord: [propagatedX, propagatedY], passedPp: passedPp, mainRoad: mainRoad, mainType: mainType, pathType: pathType)
-                                        
-                                        xyFromHead[0] = calculatedXyd[0]
-                                        xyFromHead[1] = calculatedXyd[1]
-                                        xydArray.append(calculatedXyd)
-                                        distanceSum += calculatedXyd[2]
-                                        trajectoryFromHead.append(xyFromHead)
-                                        passedPp.append(xyFromHead)
-                                    } else {
-                                        let propagatedX = xyFromHead[0] + uvLength[i]*cos(headAngle*D2R)
-                                        let propagatedY = xyFromHead[1] + uvLength[i]*sin(headAngle*D2R)
-                                        let calculatedXyd = extendedCalDistacneFromNearestPp(coord: [propagatedX, propagatedY], passedPp: passedPp, mainRoad: mainRoad, mainType: mainType, pathType: pathType)
-                                        
-                                        xyFromHead[0] = propagatedX
-                                        xyFromHead[1] = propagatedY
-                                        xydArray.append(calculatedXyd)
-                                        distanceSum += calculatedXyd[2]
-                                        trajectoryFromHead.append(xyFromHead)
-                                        passedPp.append(xyFromHead)
-                                    }
-                                } else {
-                                    let propagatedX = xyFromHead[0] + uvLength[i]*cos(headAngle*D2R)
-                                    let propagatedY = xyFromHead[1] + uvLength[i]*sin(headAngle*D2R)
-                                    let calculatedXyd = extendedCalDistacneFromNearestPp(coord: [propagatedX, propagatedY], passedPp: passedPp, mainRoad: mainRoad, mainType: mainType, pathType: pathType)
-                                    
-                                    xyFromHead[0] = calculatedXyd[0]
-                                    xyFromHead[1] = calculatedXyd[1]
-                                    xydArray.append(calculatedXyd)
-                                    distanceSum += calculatedXyd[2]
-                                    trajectoryFromHead.append(xyFromHead)
-                                    passedPp.append(xyFromHead)
-                                }
-                            }
-                            
-                            ppXydArray.append([xPath, yPath, distanceSum])
-                            
-                            if (minDistanceCoord.isEmpty) {
-                                minDistanceCoord = [xPath, yPath, distanceSum]
-                                minTrajectory = trajectoryFromHead
-                                minTrajectoryOriginal = trajectoryOriginal
-                            } else {
-                                let distanceCurrent = distanceSum
-                                let distancePast = minDistanceCoord[2]
-                                if (distanceCurrent < distancePast) {
-                                    minDistanceCoord = [xPath, yPath, distanceSum]
-                                    minTrajectory = trajectoryFromHead
-                                    minTrajectoryOriginal = trajectoryOriginal
-                                }
-                            }
-                        }
-                    }
-
-                    if (!minDistanceCoord.isEmpty) {
-                        isSuccess = true
-                        xyd = minDistanceCoord
-                    }
-                }
-            }
-        }
-        
-        matchedTraj = MatchedTraj(isSuccess: isSuccess, xyd: xyd, minTrajectory: minTrajectory, minTrajectoryOriginal: minTrajectoryOriginal)
-        
-        return matchedTraj
-    }
-    
-    public func extractMinSuccessfulMatchedTraj(_ matchedTrajs: [MatchedTraj]) -> MatchedTraj? {
-        var minSuccessfulMatchedTraj: MatchedTraj? = nil
-        var minSecondIndexValue: Double = Double.infinity
-
-        for matchedTraj in matchedTrajs {
-            if matchedTraj.isSuccess && matchedTraj.xyd.count > 1 {
-                let secondIndexValue = matchedTraj.xyd[1]
-                if secondIndexValue < minSecondIndexValue {
-                    minSecondIndexValue = secondIndexValue
-                    minSuccessfulMatchedTraj = matchedTraj
-                }
-            }
-        }
-
-        return minSuccessfulMatchedTraj
-    }
-
-    
-    func getPathMatchingHeadings(building: String, level: String, x: Double, y: Double, heading: Double, RANGE: Double, mode: String) -> [Double] {
-        var headings: [Double] = []
-        let levelCopy: String = removeLevelDirectionString(levelName: level)
-        let key: String = "\(building)_\(levelCopy)"
-        
-        if (!(building.isEmpty) && !(level.isEmpty)) {
-            guard let mainType: [Int] = self.PathType[key] else {
-                return headings
-            }
-            
-            guard let mainRoad: [[Double]] = self.PathPoint[key] else {
-                return headings
-            }
-            
-            guard let mainHeading: [String] = self.PathHeading[key] else {
-                return headings
-            }
-            
-            if (!mainRoad.isEmpty) {
-                let roadX = mainRoad[0]
-                let roadY = mainRoad[1]
-                
-                let xMin = x - RANGE
-                let xMax = x + RANGE
-                let yMin = y - RANGE
-                let yMax = y + RANGE
-                
-                for i in 0..<roadX.count {
-                    let xPath = roadX[i]
-                    let yPath = roadY[i]
-                    
-                    let pathType = mainType[i]
-                    
-                    if (mode == "dr") {
-                        if (pathType != 1) {
-                            continue
-                        }
-                    }
-                    
-                    if (xPath >= xMin && xPath <= xMax) {
-                        if (yPath >= yMin && yPath <= yMax) {
-                            let headingArray = mainHeading[i]
-                            if (!headingArray.isEmpty) {
-                                let headingData = headingArray.components(separatedBy: ",")
-                                for j in 0..<headingData.count {
-                                    if (!headingData[j].isEmpty) {
-                                        let value = Double(headingData[j])!
-                                        if (!headings.contains(value)) {
-                                            headings.append(value)
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-        return headings
-    }
 
     func checkHeadingCorrection(buffer: [Double]) -> Bool {
         if (buffer.count >= HEADING_BUFFER_SIZE) {
@@ -5808,57 +4741,16 @@ public class ServiceManager: Observation {
         timeUpdatePosition.x = timeUpdatePosition.x + dx
         timeUpdatePosition.y = timeUpdatePosition.y + dy
         timeUpdatePosition.heading = updateHeading
-        self.preTuMmHeading = compensateHeading(heading: updateHeading)
         
         var timeUpdateCopy = timeUpdatePosition
         let levelName = removeLevelDirectionString(levelName: timeUpdateOutput.level_name)
         let compensatedHeading = compensateHeading(heading: timeUpdateCopy.heading)
         if (runMode != "pdr") {
-//            let isDrStraight: Bool = isDrBufferStraight(drBuffer: drBuffer)
-//            if ((self.unitDrInfoIndex%2) == 0 && !isDrStraight) {
-//                let drBufferForPathMatching = Array(drBuffer.suffix(DR_BUFFER_SIZE_FOR_STRAIGHT))
-//                let pathTrajMatchingResult = self.extendedPathTrajectoryMatching(building: timeUpdateOutput.building_name, level: levelName, x: timeUpdateCopy.x, y: timeUpdateCopy.y, heading: compensatedHeading, pastResult: self.jupiterResult, drBuffer: drBufferForPathMatching, HEADING_RANGE: HEADING_RANGE, pathType: 1, mode: self.runMode)
-//                if (pathTrajMatchingResult.isSuccess) {
-//                    timeUpdatePosition.x = timeUpdatePosition.x*0.5 + pathTrajMatchingResult.xyd[0]*0.5
-//                    timeUpdatePosition.y = timeUpdatePosition.y*0.5 + pathTrajMatchingResult.xyd[1]*0.5
-//                    displayOutput.trajectoryPm = pathTrajMatchingResult.minTrajectory
-//                    displayOutput.trajectoryOg = pathTrajMatchingResult.minTrajectoryOriginal
-//                } else {
-//                    displayOutput.trajectoryPm = [[0,0]]
-//                    displayOutput.trajectoryOg = [[0,0]]
-//                }
-//            } else {
-//                var correctedTuCopy = (true, [timeUpdateCopy.x, timeUpdateCopy.y, timeUpdateCopy.heading])
-//                let pathMatchingResult = self.pathMatching(building: timeUpdateOutput.building_name, level: levelName, x: timeUpdateCopy.x, y: timeUpdateCopy.y, heading: compensatedHeading, tuXY: [0,0], isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: true, pathType: 1)
-//
-//                correctedTuCopy.0 = pathMatchingResult.isSuccess
-//                correctedTuCopy.1 = pathMatchingResult.xyh
-//                correctedTuCopy.1[2] = compensateHeading(heading: correctedTuCopy.1[2])
-//                if (correctedTuCopy.0) {
-//                    timeUpdateCopy.x = correctedTuCopy.1[0]
-//                    timeUpdateCopy.y = correctedTuCopy.1[1]
-//                    if (isNeedHeadingCorrection) {
-//                        timeUpdateCopy.heading = correctedTuCopy.1[2]
-//                    }
-//                    timeUpdatePosition = timeUpdateCopy
-//                } else {
-//                    correctedTuCopy.0 = pathMatchingResult.0
-//                    correctedTuCopy.1 = pathMatchingResult.1
-//
-//                    timeUpdateCopy.x = correctedTuCopy.1[0]
-//                    timeUpdateCopy.y = correctedTuCopy.1[1]
-//                    timeUpdatePosition = timeUpdateCopy
-//                }
-//
-//                displayOutput.trajectoryPm = [[0,0]]
-//                displayOutput.trajectoryOg = [[0,0]]
-//            }
-            
-            var correctedTuCopy = (true, [timeUpdateCopy.x, timeUpdateCopy.y, timeUpdateCopy.heading])
-            let pathMatchingResult = self.pathMatching(building: timeUpdateOutput.building_name, level: levelName, x: timeUpdateCopy.x, y: timeUpdateCopy.y, heading: compensatedHeading, tuXY: [0,0], isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: true, pathType: 1)
+            var correctedTuCopy = (true, [timeUpdateCopy.x, timeUpdateCopy.y, timeUpdateCopy.heading, 1.0])
+            let pathMatchingResult = pmCalculator.pathMatching(building: timeUpdateOutput.building_name, level: levelName, x: timeUpdateCopy.x, y: timeUpdateCopy.y, heading: compensatedHeading, isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: true, pathType: 1, range: SQUARE_RANGE)
 
             correctedTuCopy.0 = pathMatchingResult.isSuccess
-            correctedTuCopy.1 = pathMatchingResult.xyh
+            correctedTuCopy.1 = pathMatchingResult.xyhs
             correctedTuCopy.1[2] = compensateHeading(heading: correctedTuCopy.1[2])
             if (correctedTuCopy.0) {
                 timeUpdateCopy.x = correctedTuCopy.1[0]
@@ -5915,15 +4807,15 @@ public class ServiceManager: Observation {
         serverOutputHatCopy.absolute_heading = compensateHeading(heading: serverOutputHatCopy.absolute_heading)
         
         // ServerOutputHat을 맵매칭
-        var serverOutputHatCopyMm = (true, [serverOutputHatCopy.x, serverOutputHatCopy.y, serverOutputHatCopy.absolute_heading])
+        var serverOutputHatCopyMm = (true, [serverOutputHatCopy.x, serverOutputHatCopy.y, serverOutputHatCopy.absolute_heading, 1.0])
         if (self.runMode == "pdr") {
-            let pathMatchingResult = self.pathMatching(building: serverOutputHatCopy.building_name, level: serverOutputHatCopy.level_name, x: serverOutputHatCopy.x, y: serverOutputHatCopy.y, heading: serverOutputHatCopy.absolute_heading, tuXY: [0,0], isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: false, pathType: 0)
+            let pathMatchingResult = pmCalculator.pathMatching(building: serverOutputHatCopy.building_name, level: serverOutputHatCopy.level_name, x: serverOutputHatCopy.x, y: serverOutputHatCopy.y, heading: serverOutputHatCopy.absolute_heading, isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: false, pathType: 0, range: SQUARE_RANGE)
             serverOutputHatCopyMm.0 = pathMatchingResult.isSuccess
-            serverOutputHatCopyMm.1 = pathMatchingResult.xyh
+            serverOutputHatCopyMm.1 = pathMatchingResult.xyhs
         } else {
-            let pathMatchingResult = self.pathMatching(building: serverOutputHatCopy.building_name, level: serverOutputHatCopy.level_name, x: serverOutputHatCopy.x, y: serverOutputHatCopy.y, heading: serverOutputHatCopy.absolute_heading, tuXY: [0,0], isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: true, pathType: 1)
+            let pathMatchingResult = pmCalculator.pathMatching(building: serverOutputHatCopy.building_name, level: serverOutputHatCopy.level_name, x: serverOutputHatCopy.x, y: serverOutputHatCopy.y, heading: serverOutputHatCopy.absolute_heading, isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: true, pathType: 1, range: SQUARE_RANGE)
             serverOutputHatCopyMm.0 = pathMatchingResult.isSuccess
-            serverOutputHatCopyMm.1 = pathMatchingResult.xyh
+            serverOutputHatCopyMm.1 = pathMatchingResult.xyhs
         }
         serverOutputHatCopyMm.1[2] = compensateHeading(heading: serverOutputHatCopyMm.1[2])
         
@@ -5963,15 +4855,15 @@ public class ServiceManager: Observation {
         kalmanP -= kalmanK * kalmanP
         headingKalmanP -= headingKalmanK * headingKalmanP
         
-        var measurementOutputCorrected = (true, [measurementOutput.x, measurementOutput.y, updateHeading])
+        var measurementOutputCorrected = (true, [measurementOutput.x, measurementOutput.y, updateHeading, 1.0])
         if (self.runMode == "pdr") {
-            let pathMatchingResult = self.pathMatching(building: measurementOutput.building_name, level: measurementOutput.level_name, x: measurementOutput.x, y: measurementOutput.y, heading: updateHeading, tuXY: [0,0], isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: false, pathType: 0)
+            let pathMatchingResult = pmCalculator.pathMatching(building: measurementOutput.building_name, level: measurementOutput.level_name, x: measurementOutput.x, y: measurementOutput.y, heading: updateHeading, isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: false, pathType: 0, range: SQUARE_RANGE)
             measurementOutputCorrected.0 = pathMatchingResult.isSuccess
-            measurementOutputCorrected.1 = pathMatchingResult.xyh
+            measurementOutputCorrected.1 = pathMatchingResult.xyhs
         } else {
-            let pathMatchingResult = self.pathMatching(building: measurementOutput.building_name, level: measurementOutput.level_name, x: measurementOutput.x, y: measurementOutput.y, heading: updateHeading, tuXY: [0,0], isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: true, pathType: 1)
+            let pathMatchingResult = pmCalculator.pathMatching(building: measurementOutput.building_name, level: measurementOutput.level_name, x: measurementOutput.x, y: measurementOutput.y, heading: updateHeading, isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: true, pathType: 1, range: SQUARE_RANGE)
             measurementOutputCorrected.0 = pathMatchingResult.isSuccess
-            measurementOutputCorrected.1 = pathMatchingResult.xyh
+            measurementOutputCorrected.1 = pathMatchingResult.xyhs
         }
         measurementOutputCorrected.1[2] = compensateHeading(heading: measurementOutputCorrected.1[2])
         
@@ -5995,11 +4887,11 @@ public class ServiceManager: Observation {
                 if (propagationResult.0) {
                     var propagatedResult: [Double] = [originalResult[0]+propagationValues[0] , originalResult[1]+propagationValues[1], originalResult[2]+propagationValues[2]]
                     if (self.runMode == "pdr") {
-                        let pathMatchingResult = self.pathMatching(building: serverResult.building_name, level: serverResult.level_name, x: propagatedResult[0], y: propagatedResult[1], heading: propagatedResult[2], tuXY: [0,0], isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: false, pathType: 0)
-                        propagatedResult = pathMatchingResult.xyh
+                        let pathMatchingResult = pmCalculator.pathMatching(building: serverResult.building_name, level: serverResult.level_name, x: propagatedResult[0], y: propagatedResult[1], heading: propagatedResult[2], isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: false, pathType: 0, range: SQUARE_RANGE)
+                        propagatedResult = pathMatchingResult.xyhs
                     } else {
-                        let pathMatchingResult = self.pathMatching(building: serverResult.building_name, level: serverResult.level_name, x: propagatedResult[0], y: propagatedResult[1], heading: propagatedResult[2], tuXY: [0,0], isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: true, pathType: 1)
-                        propagatedResult = pathMatchingResult.xyh
+                        let pathMatchingResult = pmCalculator.pathMatching(building: serverResult.building_name, level: serverResult.level_name, x: propagatedResult[0], y: propagatedResult[1], heading: propagatedResult[2], isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: true, pathType: 1, range: SQUARE_RANGE)
+                        propagatedResult = pathMatchingResult.xyhs
                     }
                     propagatedResult[2] = compensateHeading(heading: propagatedResult[2])
                     measurementOutputCorrected.1 = propagatedResult
@@ -6043,15 +4935,15 @@ public class ServiceManager: Observation {
                 saveKalmanParam()
             }
         } else {
-            var measurementOutputCorrected = (true, [measurementOutput.x, measurementOutput.y, updateHeading])
+            var measurementOutputCorrected = (true, [measurementOutput.x, measurementOutput.y, updateHeading, 1.0])
             if (self.runMode == "pdr") {
-                let pathMatchingResult = self.pathMatching(building: measurementOutput.building_name, level: measurementOutput.level_name, x: measurementOutput.x, y: measurementOutput.y, heading: updateHeading, tuXY: [0,0], isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: false, pathType: 0)
+                let pathMatchingResult = pmCalculator.pathMatching(building: measurementOutput.building_name, level: measurementOutput.level_name, x: measurementOutput.x, y: measurementOutput.y, heading: updateHeading, isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: false, pathType: 0, range: SQUARE_RANGE)
                 measurementOutputCorrected.0 = pathMatchingResult.isSuccess
-                measurementOutputCorrected.1 = pathMatchingResult.xyh
+                measurementOutputCorrected.1 = pathMatchingResult.xyhs
             } else {
-                let pathMatchingResult = self.pathMatching(building: measurementOutput.building_name, level: measurementOutput.level_name, x: measurementOutput.x, y: measurementOutput.y, heading: updateHeading, tuXY: [0,0], isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: false, pathType: 1)
+                let pathMatchingResult = pmCalculator.pathMatching(building: measurementOutput.building_name, level: measurementOutput.level_name, x: measurementOutput.x, y: measurementOutput.y, heading: updateHeading, isPast: false, HEADING_RANGE: HEADING_RANGE, isUseHeading: false, pathType: 1, range: SQUARE_RANGE)
                 measurementOutputCorrected.0 = pathMatchingResult.isSuccess
-                measurementOutputCorrected.1 = pathMatchingResult.xyh
+                measurementOutputCorrected.1 = pathMatchingResult.xyhs
             }
             measurementOutputCorrected.1[2] = compensateHeading(heading: measurementOutputCorrected.1[2])
             
@@ -6101,30 +4993,5 @@ public class ServiceManager: Observation {
         } else {
             self.BLE_VALID_TIME = 1500
         }
-    }
-    
-    func findNetworkBadEntrance(bleAvg: [String: Double], bias: Int) -> (Bool, FineLocationTrackingFromServer) {
-        var isInNetworkBadEntrance: Bool = false
-        var entrance = FineLocationTrackingFromServer()
-        
-        let networkBadEntranceWards = ["TJ-00CB-00000386-0000"]
-        for (key, value) in bleAvg {
-            if networkBadEntranceWards.contains(key) {
-                let rssi = value + Double(bias)
-                if (rssi >= -82.0) {
-                    isInNetworkBadEntrance = true
-                    
-                    entrance.building_name = "COEX"
-                    entrance.level_name = "B0"
-                    entrance.x = 270
-                    entrance.y = 10
-                    entrance.absolute_heading = 270
-                    
-                    return (isInNetworkBadEntrance, entrance)
-                }
-            }
-        }
-        
-        return (isInNetworkBadEntrance, entrance)
     }
 }
